Create CQL filter strings from sf objects

Convenience wrapper to convert sf objects and geometric operations into CQL filter strings which can then be supplied to the ... argument in bcdc_get_geodata. The sf object is automatically converted in a bounding box to reduce the complexity of the Web Service call. Subsequent in-memory filtering may be need to achieve exact results.

bcdc_cql_string(x, geometry_predicates, pattern = NULL,
  distance = NULL, units = NULL, coords = NULL, crs = NULL)

Arguments

x
object of class sf, sfc or sfg

geometry_predicates
Geometry predicates that allow for spatial filtering. bcbdc_cql_string accepts the following geometric predicates: EQUALS, DISJOINT, INTERSECTS, TOUCHES, CROSSES, WITHIN, CONTAINS, OVERLAPS, RELATE, DWITHIN, BEYOND, BBOX.

pattern
spatial relationship specified by a DE-9IM matrix pattern. A DE-9IM pattern is a string of length 9 specified using the characters *TF012. Example: '1*T***T**'

distance
numeric value for distance tolerance

units
units that distance is specified in. One of "feet", "meters", "statute miles", "nautical miles", "kilometers"

coords
the coordinates of the bounding box.

crs
(Optional) A string containing an SRS code (For example, 'EPSG:1234'. The default is to use the CRS of the queried layer)

Details

There are wrapper functions for each of the geometry predicates, which you may find more convenient than calling this function directly.

See also

Examples

# NOT RUN {
airports <- bcdc_get_geodata("bc-airports") %>% collect()
bcdc_cql_string(airports, "DWITHIN")
# }