Download and read a resource from a B.C. Data Catalogue record

Download and read a resource from a B.C. Data Catalogue record

bcdc_get_data(record, resource = NULL, verbose = TRUE, ...)

Arguments

record
either a bcdc_record object (from the result of bcdc_get_record()), a character string denoting the name or ID of a resource (or the URL) or a BC Geographic Warehouse (BCGW) name.

It is advised to use the permanent ID for a record or the BCGW name rather than the human-readable name to guard against future name changes of the record. If you use the human-readable name a warning will be issued once per session. You can silence these warnings altogether by setting an option: options("silence_named_get_data_warning" = TRUE) - which you can set in your .Rprofile file so the option persists across sessions.

resource
optional argument used when there are multiple data files within the same record. See examples.

verbose
When more than one resource is available for a record, should extra information about those resources be printed to the console? Default TRUE

...
arguments passed to other functions. Tabular data is passed to a function to handle the import based on the file extension. bcdc_read_functions() provides details on which functions handle the data import. You can then use this information to look at the help pages of those functions. See the examples for a workflow that illustrates this process. For spatial Web Service data the ... arguments are passed to bcdc_query_geodata().

Value

An object of a type relevant to the resource (usually a tibble or an sf object)

Examples

# \donttest{
# Using the record and resource ID:
bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8',
              resource = '4d0377d9-e8a1-429b-824f-0ce8f363512c')

#> Simple feature collection with 455 features and 41 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 406543.7 ymin: 367957.6 xmax: 1796645 ymax: 1689146
#> epsg (SRID):    3005
#> proj4string:    +proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 455 x 42
#>    id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>  * <chr> <chr>            <chr>            <chr>            <chr>           
#>  1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  4 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  5 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  6 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  7 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  8 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  9 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 10 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 445 more rows, and 37 more variables: SOURCE_DATA_ID <chr>,
#> #   SUPPLIED_SOURCE_ID_IND <chr>, AIRPORT_NAME <chr>, DESCRIPTION <chr>,
#> #   PHYSICAL_ADDRESS <chr>, ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>,
#> #   POSTAL_CODE <chr>, LOCALITY <chr>, CONTACT_PHONE <chr>,
#> #   CONTACT_EMAIL <chr>, CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>,
#> #   LATITUDE <dbl>, LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <dbl>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

bcdc_get_data('1d21922b-ec4f-42e5-8f6b-bf320a286157')

#> Reading the data using the read_csv function from the readr package.

#> Parsed with column specification:
#> cols(
#>   CITY = col_character(),
#>   NAME = col_character(),
#>   ADDRESS = col_character(),
#>   POSTALCODE = col_character()
#> )

#> # A tibble: 357 x 4
#>    CITY        NAME                    ADDRESS                        POSTALCODE
#>    <chr>       <chr>                   <chr>                          <chr>     
#>  1 PRINCE GEO… Pacific Western Brewin… 641 NORTH NECHAKO RD           V2K4M4    
#>  2 OSOYOOS     Bordertown Vineyards    9140 92ND AVENUE               V0H1V2    
#>  3 CHILLIWACK  Whispering Horse Winery 43721 VEDDER MTN. ROAD         V2R4C5    
#>  4 KELOWNA     Okanaganvilla Estate W… 3240 POOLEY ROAD               V1W4G7    
#>  5 OLIVER      Church & State Wines (… 31120 87TH STREET              V0H1T0    
#>  6 KEREMEOS    Corcelettes Estate Win… 2582 UPPER BENCH ROAD          V0X1N4    
#>  7 DUNCAN      Blue Grouse Vineyards   4365 BLUE GROUSE ROAD AND 218… V9L6M3    
#>  8 VANCOUVER   City Side Winery        328 WEST 2ND AVENUE            V5Y1C8    
#>  9 WEST KELOW… Volcanic Hills Estate … 2845 BOUCHERIE ROAD            V1Z2G6    
#> 10 OLIVER      Orchard Hill Estate Ci… 3480 FRUITVALE WAY             V0H1T0    
#> # … with 347 more rows

# Using a `bcdc_record` object obtained from `bcdc_get_record`:
record <- bcdc_get_record('1d21922b-ec4f-42e5-8f6b-bf320a286157')
bcdc_get_data(record)

#> Reading the data using the read_csv function from the readr package.

#> Parsed with column specification:
#> cols(
#>   CITY = col_character(),
#>   NAME = col_character(),
#>   ADDRESS = col_character(),
#>   POSTALCODE = col_character()
#> )

#> # A tibble: 357 x 4
#>    CITY        NAME                    ADDRESS                        POSTALCODE
#>    <chr>       <chr>                   <chr>                          <chr>     
#>  1 PRINCE GEO… Pacific Western Brewin… 641 NORTH NECHAKO RD           V2K4M4    
#>  2 OSOYOOS     Bordertown Vineyards    9140 92ND AVENUE               V0H1V2    
#>  3 CHILLIWACK  Whispering Horse Winery 43721 VEDDER MTN. ROAD         V2R4C5    
#>  4 KELOWNA     Okanaganvilla Estate W… 3240 POOLEY ROAD               V1W4G7    
#>  5 OLIVER      Church & State Wines (… 31120 87TH STREET              V0H1T0    
#>  6 KEREMEOS    Corcelettes Estate Win… 2582 UPPER BENCH ROAD          V0X1N4    
#>  7 DUNCAN      Blue Grouse Vineyards   4365 BLUE GROUSE ROAD AND 218… V9L6M3    
#>  8 VANCOUVER   City Side Winery        328 WEST 2ND AVENUE            V5Y1C8    
#>  9 WEST KELOW… Volcanic Hills Estate … 2845 BOUCHERIE ROAD            V1Z2G6    
#> 10 OLIVER      Orchard Hill Estate Ci… 3480 FRUITVALE WAY             V0H1T0    
#> # … with 347 more rows

# Using a BCGW name
bcdc_get_data("WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW")

#> Simple feature collection with 455 features and 41 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 406543.7 ymin: 367957.6 xmax: 1796645 ymax: 1689146
#> epsg (SRID):    3005
#> proj4string:    +proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 455 x 42
#>    id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>  * <chr> <chr>            <chr>            <chr>            <chr>           
#>  1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  4 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  5 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  6 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  7 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  8 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  9 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 10 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 445 more rows, and 37 more variables: SOURCE_DATA_ID <chr>,
#> #   SUPPLIED_SOURCE_ID_IND <chr>, AIRPORT_NAME <chr>, DESCRIPTION <chr>,
#> #   PHYSICAL_ADDRESS <chr>, ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>,
#> #   POSTAL_CODE <chr>, LOCALITY <chr>, CONTACT_PHONE <chr>,
#> #   CONTACT_EMAIL <chr>, CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>,
#> #   LATITUDE <dbl>, LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <dbl>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

## Example of correcting import problems

## Some initial problems reading in the data
bcdc_get_data('d7e6c8c7-052f-4f06-b178-74c02c243ea4')

#> Reading the data using the read_xlsx function from the readxl package.

#> New names:
#> * `` -> ...2
#> * `` -> ...3
#> * `` -> ...4
#> * `` -> ...5
#> * `` -> ...6
#> * … and 5 more problems

#> # A tibble: 901 x 11
#>    `Regional Distri… ...2  ...3  ...4  ...5  ...6  ...7  ...8  ...9  ...10 ...11
#>    <chr>             <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 RD Code           Regi… Gend… Year  Age … 15-19 20-24 25-64 65-84 85+   All …
#>  2 1000              East… Male  2018  4876  1708  1814  18244 5581  510   32733
#>  3 1000              East… Male  2019  4895  1655  1839  18152 5808  548   32897
#>  4 1000              East… Male  2020  4912  1614  1816  18085 6050  570   33047
#>  5 1000              East… Male  2021  4938  1585  1841  17914 6326  608   33212
#>  6 1000              East… Male  2022  4906  1616  1725  17937 6570  613   33367
#>  7 1000              East… Male  2023  4922  1594  1661  17892 6839  602   33510
#>  8 1000              East… Male  2024  4929  1593  1600  17802 7078  631   33633
#>  9 1000              East… Male  2025  4903  1622  1560  17666 7287  707   33745
#> 10 1000              East… Male  2026  4853  1680  1534  17518 7569  705   33859
#> # … with 891 more rows

## From bcdc_get_record we realize that the data is in xlsx format
bcdc_get_record('d7e6c8c7-052f-4f06-b178-74c02c243ea4')

#> B.C. Data Catalogue Record: Local Government Population and Household Projections (2018-2027) 
#> 
#> Name: local-government-population-and-household-projections-2018-2027- (ID: d7e6c8c7-052f-4f06-b178-74c02c243ea4 )
#> Permalink: https://catalogue.data.gov.bc.ca/dataset/d7e6c8c7-052f-4f06-b178-74c02c243ea4
#> Sector: Economy
#> Licence: Access Only
#> Type: Dataset
#> Last Updated: 2019-07-17 
#> 
#> Description: This dataset contains population and household projections for 2018-2027 including:
#> overall population, gender distribution, age distribution, median age, average age,
#> population growth rates, number of households, average number of people per
#> household.  
#> 
#> Resources: (1)
#> # A tibble: 1 x 8
#>   name      url        id     format ext   package_id  location bcdata_available
#>   <chr>     <chr>      <chr>  <chr>  <chr> <chr>       <chr>    <lgl>           
#> 1 Populati… https://c… d39bd… xlsx   xlsx  d7e6c8c7-0… catalog… TRUE            
#> 
#> You can access the 'Resources' data frame using bcdc_tidy_resources()
#> 
#> 

## bcdc_read_functions let's us know that bcdata
## uses readxl::read_excel to import xlsx files
bcdc_read_functions()

#> # A tibble: 11 x 3
#>    format  package fun      
#>    <chr>   <chr>   <chr>    
#>  1 kml     sf      read_sf  
#>  2 geojson sf      read_sf  
#>  3 gpkg    sf      read_sf  
#>  4 gdb     sf      read_sf  
#>  5 fgdb    sf      read_sf  
#>  6 shp     sf      read_sf  
#>  7 csv     readr   read_csv 
#>  8 txt     readr   read_tsv 
#>  9 tsv     readr   read_tsv 
#> 10 xlsx    readxl  read_xlsx
#> 11 xls     readxl  read_xls 

## If you read the help page for readxl::read_excel,
## it seems likely that we need to skip the first row:
bcdc_get_data('d7e6c8c7-052f-4f06-b178-74c02c243ea4', skip = 1)

#> Reading the data using the read_xlsx function from the readxl package.

#> # A tibble: 900 x 11
#>    `RD Code` `Regional Distr… Gender  Year `Age 0-14` `15-19` `20-24` `25-64`
#>        <dbl> <chr>            <chr>  <dbl>      <dbl>   <dbl>   <dbl>   <dbl>
#>  1      1000 East Kootenay    Male    2018       4876    1708    1814   18244
#>  2      1000 East Kootenay    Male    2019       4895    1655    1839   18152
#>  3      1000 East Kootenay    Male    2020       4912    1614    1816   18085
#>  4      1000 East Kootenay    Male    2021       4938    1585    1841   17914
#>  5      1000 East Kootenay    Male    2022       4906    1616    1725   17937
#>  6      1000 East Kootenay    Male    2023       4922    1594    1661   17892
#>  7      1000 East Kootenay    Male    2024       4929    1593    1600   17802
#>  8      1000 East Kootenay    Male    2025       4903    1622    1560   17666
#>  9      1000 East Kootenay    Male    2026       4853    1680    1534   17518
#> 10      1000 East Kootenay    Male    2027       4825    1658    1563   17358
#> # … with 890 more rows, and 3 more variables: `65-84` <dbl>, `85+` <dbl>, `All
#> #   Ages` <dbl>

# }