Show a single B.C. Data Catalogue record

Show a single B.C. Data Catalogue record

bcdc_get_record(id)

Arguments

id
the human-readable name, permalink ID, or URL of the record.

It is advised to use the permanent ID for a record rather than the human-readable name to guard against future name changes of the record. If you use the human-readable name a warning will be issued once per session. You can silence these warnings altogether by setting an option: options("silence_named_get_record_warning" = TRUE) - which you can put in your .Rprofile file so the option persists across sessions.

Value

A list containing the metadata for the record

Examples

# \donttest{
bcdc_get_record("https://catalogue.data.gov.bc.ca/dataset/bc-airports")

#> B.C. Data Catalogue Record: BC Airports 
#> 
#> Name: bc-airports (ID: 76b1b7a3-2112-4444-857a-afccf7b20da8 )
#> Permalink: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8
#> Sector: Natural Resources
#> Licence: Open Government Licence - British Columbia
#> Type: Geographic
#> Last Updated: 2019-12-17 
#> 
#> Description: BC Airports identifies locations where aircraft may take-off and land. No guarantee
#> is given that an identified point will be maintained to sufficient standards for
#> landing and take-off of any/all aircraft.  It includes airports, aerodromes, water
#> aerodromes, heliports, and airstrips. 
#> 
#> Resources: (4)
#> # A tibble: 4 x 8
#>   name     url        id      format ext   package_id  location bcdata_available
#>   <chr>    <chr>      <chr>   <chr>  <chr> <chr>       <chr>    <lgl>           
#> 1 BC Geog… https://c… 604c8b… other  ""    76b1b7a3-2… bcgeogr… FALSE           
#> 2 BC_Airp… https://c… fcccba… xlsx   xlsx  76b1b7a3-2… catalog… TRUE            
#> 3 WMS get… https://o… 4d0377… wms    ""    76b1b7a3-2… bcgeogr… TRUE            
#> 4 KML Net… http://op… 5b9f22… kml    kml   76b1b7a3-2… bcgeogr… FALSE           
#> 
#> You can access the 'Resources' data frame using bcdc_tidy_resources()
#> 
#> 

bcdc_get_record("bc-airports")

#> B.C. Data Catalogue Record: BC Airports 
#> 
#> Name: bc-airports (ID: 76b1b7a3-2112-4444-857a-afccf7b20da8 )
#> Permalink: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8
#> Sector: Natural Resources
#> Licence: Open Government Licence - British Columbia
#> Type: Geographic
#> Last Updated: 2019-12-17 
#> 
#> Description: BC Airports identifies locations where aircraft may take-off and land. No guarantee
#> is given that an identified point will be maintained to sufficient standards for
#> landing and take-off of any/all aircraft.  It includes airports, aerodromes, water
#> aerodromes, heliports, and airstrips. 
#> 
#> Resources: (4)
#> # A tibble: 4 x 8
#>   name     url        id      format ext   package_id  location bcdata_available
#>   <chr>    <chr>      <chr>   <chr>  <chr> <chr>       <chr>    <lgl>           
#> 1 BC Geog… https://c… 604c8b… other  ""    76b1b7a3-2… bcgeogr… FALSE           
#> 2 BC_Airp… https://c… fcccba… xlsx   xlsx  76b1b7a3-2… catalog… TRUE            
#> 3 WMS get… https://o… 4d0377… wms    ""    76b1b7a3-2… bcgeogr… TRUE            
#> 4 KML Net… http://op… 5b9f22… kml    kml   76b1b7a3-2… bcgeogr… FALSE           
#> 
#> You can access the 'Resources' data frame using bcdc_tidy_resources()
#> 
#> 

bcdc_get_record("https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8")

#> B.C. Data Catalogue Record: BC Airports 
#> 
#> Name: bc-airports (ID: 76b1b7a3-2112-4444-857a-afccf7b20da8 )
#> Permalink: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8
#> Sector: Natural Resources
#> Licence: Open Government Licence - British Columbia
#> Type: Geographic
#> Last Updated: 2019-12-17 
#> 
#> Description: BC Airports identifies locations where aircraft may take-off and land. No guarantee
#> is given that an identified point will be maintained to sufficient standards for
#> landing and take-off of any/all aircraft.  It includes airports, aerodromes, water
#> aerodromes, heliports, and airstrips. 
#> 
#> Resources: (4)
#> # A tibble: 4 x 8
#>   name     url        id      format ext   package_id  location bcdata_available
#>   <chr>    <chr>      <chr>   <chr>  <chr> <chr>       <chr>    <lgl>           
#> 1 BC Geog… https://c… 604c8b… other  ""    76b1b7a3-2… bcgeogr… FALSE           
#> 2 BC_Airp… https://c… fcccba… xlsx   xlsx  76b1b7a3-2… catalog… TRUE            
#> 3 WMS get… https://o… 4d0377… wms    ""    76b1b7a3-2… bcgeogr… TRUE            
#> 4 KML Net… http://op… 5b9f22… kml    kml   76b1b7a3-2… bcgeogr… FALSE           
#> 
#> You can access the 'Resources' data frame using bcdc_tidy_resources()
#> 
#> 

bcdc_get_record("76b1b7a3-2112-4444-857a-afccf7b20da8")

#> B.C. Data Catalogue Record: BC Airports 
#> 
#> Name: bc-airports (ID: 76b1b7a3-2112-4444-857a-afccf7b20da8 )
#> Permalink: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8
#> Sector: Natural Resources
#> Licence: Open Government Licence - British Columbia
#> Type: Geographic
#> Last Updated: 2019-12-17 
#> 
#> Description: BC Airports identifies locations where aircraft may take-off and land. No guarantee
#> is given that an identified point will be maintained to sufficient standards for
#> landing and take-off of any/all aircraft.  It includes airports, aerodromes, water
#> aerodromes, heliports, and airstrips. 
#> 
#> Resources: (4)
#> # A tibble: 4 x 8
#>   name     url        id      format ext   package_id  location bcdata_available
#>   <chr>    <chr>      <chr>   <chr>  <chr> <chr>       <chr>    <lgl>           
#> 1 BC Geog… https://c… 604c8b… other  ""    76b1b7a3-2… bcgeogr… FALSE           
#> 2 BC_Airp… https://c… fcccba… xlsx   xlsx  76b1b7a3-2… catalog… TRUE            
#> 3 WMS get… https://o… 4d0377… wms    ""    76b1b7a3-2… bcgeogr… TRUE            
#> 4 KML Net… http://op… 5b9f22… kml    kml   76b1b7a3-2… bcgeogr… FALSE           
#> 
#> You can access the 'Resources' data frame using bcdc_tidy_resources()
#> 
#> 

# }