Retrieve group information for B.C. Data Catalogue

Returns a tibble of groups or records. Groups can be viewed here: https://catalogue.data.gov.bc.ca/group or accessed directly from R using bcdc_list_groups

bcdc_list_groups()

bcdc_list_group_records(group)

Arguments

group
Name of the group

Functions

bcdc_list_groups:

Examples

# \donttest{
bcdc_list_group_records('environmental-reporting-bc')

#> Group Description: Environmental Reporting BC
#>  examines the state and trends of environmental measures across the
#>  province. We provide access to scientific data and information about
#>  our environment and how it relates to British Columbians. See for
#>  [More
#>  Information](https://www2.gov.bc.ca/gov/content?id=FF80E0B985F245CEA62808414D78C41B)
#>  and the [GitHub Repo](https://github.com/bcgov/EnvReportBC).
#> Number of datasets: 32
#> # A tibble: 32 x 3
#>    title                          id               name                         
#>    <chr>                          <chr>            <chr>                        
#>  1 Indicator Summary Data: Trend… 54ec827b-3b9a-4… indicator-summary-data-trend…
#>  2 Indicator Summary Data: Amoun… 42f52a87-e0a3-4… indicator-summary-data-amoun…
#>  3 Indicator Summary Data: Trend… 18754165-1daa-4… indicator-summary-data-trend…
#>  4 Indicator Summary Data: Trend… b3369823-d130-4… indicator-summary-data-trend…
#>  5 Land Designations that Contri… e0b30537-5498-4… land-designations-that-contr…
#>  6 Indicator Summary Data: Land … 2ad14e62-e0b1-4… indicator-summary-data-land-…
#>  7 Inventory of Air Quality Byla… 8c9dbd87-3704-4… inventory-of-air-quality-byl…
#>  8 Indicator Summary Data: Chang… 86526746-40dd-4… indicator-summary-data-chang…
#>  9 Indicator Summary Data: Chang… d6f30634-a6a8-4… indicator-summary-data-chang…
#> 10 Indicator Summary Data: Chang… ad95a5c1-5e3e-4… indicator-summary-data-chang…
#> # … with 22 more rows

# }