Search the B.C. Data Catalogue

Search the B.C. Data Catalogue

bcdc_search(
  ...,
  license_id = NULL,
  download_audience = "Public",
  res_format = NULL,
  sector = NULL,
  organization = NULL,
  n = 100
)

Arguments

...
search terms

license_id
the type of license (see bcdc_search_facets("license_id")).

download_audience
download audience (see bcdc_search_facets("download_audience")). Default "Public"

res_format
format of resource (see bcdc_search_facets("res_format"))

sector
sector of government from which the data comes (see bcdc_search_facets("sector"))

organization
government organization that manages the data (see bcdc_search_facets("organization"))

n
number of results to return. Default 100

Value

A list containing the records that match the search

Examples

# \donttest{
try(
  bcdc_search("forest")
)

#> Found 275 matches. Returning the first 100.
#> To see them all, rerun the search and set the 'n' argument to 275.

#> List of B.C. Data Catalogue Records
#> Number of records: 100
#>  (Showing the top 10)
#> Titles:
#> 1: Forest Development Units (other, wms, kml)
#>  ID: 25bbc9a6-f0eb-47dd-849c-3161a6512468
#>  Name: forest-development-units
#> 2: Harvested Areas of BC (Consolidated Cutblocks) (other, fgdb, wms,
#>  kml, pdf)
#>  ID: b1b647a6-f271-42e0-9cd0-89ec24bce9f7
#>  Name: harvested-areas-of-bc-consolidated-cutblocks-
#> 3: RESULTS - Forest Cover Silviculture (other, wms, kml)
#>  ID: 258bb088-4113-47b1-b568-ce20bd64e3e3
#>  Name: results-forest-cover-silviculture
#> 4: RESULTS - Forest Cover Inventory (other, wms)
#>  ID: 56ac43a7-724a-4f01-b193-d5f9a16ef0a8
#>  Name: results-forest-cover-inventory
#> 5: Forest Tenure Managed Licence (other, wms, kml)
#>  ID: c3e96239-cdc9-4328-ac19-58fba1623ef8
#>  Name: forest-tenure-managed-licence
#> 6: Pest Infestation Polygons (other, zip)
#>  ID: 450b67bb-02d5-4526-8bc0-ac7924125a1e
#>  Name: pest-infestation-polygons
#> 7: Pest Infestation Points (other)
#>  ID: 4a83ec61-31a4-4134-9204-af47ff26ed78
#>  Name: pest-infestation-points
#> 8: Forest Tenure Harvesting Authority Polygons (other, wms, kml)
#>  ID: cff7b8f7-6897-444f-8c53-4bb93c7e9f8b
#>  Name: forest-tenure-harvesting-authority-polygons
#> 9: Generalized Forest Cover Ownership (wms, kml, other)
#>  ID: 5fc4e8ce-dd1d-44fd-af17-e0789cf65e4e
#>  Name: generalized-forest-cover-ownership
#> 10: FDU - Forest Licence Reference (other)
#>  ID: 7e986391-c740-4d1b-9294-7576af2d2998
#>  Name: fdu-forest-licence-reference
#> 
#> Access a single record by calling `bcdc_get_record(ID)` with the ID
#>  from the desired record.

try(
  bcdc_search("regional district", res_format = "fgdb")
)

#> List of B.C. Data Catalogue Records
#> Number of records: 95
#>  (Showing the top 10)
#> Titles:
#> 1: Regional District Boundaries - Road centreline aligned (other, wms,
#>  kml)
#>  ID: b66ed992-16c9-4527-b363-0ae4827d103b
#>  Name: regional-district-boundaries-road-centreline-aligned
#> 2: Regional district financial statistics (other)
#>  ID: 35fd9fe7-e937-4719-82e8-852cc699a113
#>  Name: regional-district-financial-statistics
#> 3: Regional Districts - Legally Defined Administrative Areas of BC
#>  (other, xlsx, wms, kml)
#>  ID: d1aff64e-dbfe-45a6-af97-582b7f6418b9
#>  Name: regional-districts-legally-defined-administrative-areas-of-bc
#> 4: Custom Census Reports (2016, 2011, 2006) (other, xlsx)
#>  ID: 894497d9-6901-4ed8-8655-79010e4b9b09
#>  Name: custom-census-reports-2016-2011-2006-
#> 5: CEEI Supporting Indicators Total 2009 Regional Districts (other)
#>  ID: 32966665-8e5c-4f9e-b10d-311355f2a1f0
#>  Name: ceei-supporting-indicators-total-2009-regional-districts
#> 6: CEEI Supporting Indicators Total 2006 Regional Districts (other)
#>  ID: d1be3579-82af-489b-90dd-3463c06dee58
#>  Name: ceei-supporting-indicators-total-2006-regional-districts
#> 7: CEEI Supporting Indicators Total 2001 Regional Districts (other)
#>  ID: 21486b7c-480a-4757-a7f8-8c780d156249
#>  Name: ceei-supporting-indicators-total-2001-regional-districts
#> 8: CEEI Supporting Indicators Total 1996 Regional Districts (other)
#>  ID: dd1c1225-3a4c-4181-85d2-2c07f05a91ca
#>  Name: ceei-supporting-indicators-total-1996-regional-districts
#> 9: Recreation Sites and Trails BC District Boundaries (fgdb, other,
#>  wms, kml)
#>  ID: 0ad49841-a25b-4be4-b98c-9e817e94cc48
#>  Name: recreation-sites-and-trails-bc-district-boundaries
#> 10: Forest Licensee Operating Areas (other, wms, kml)
#>  ID: 5a18c401-7c99-48fe-8a1a-b4597072d8b6
#>  Name: forest-licensee-operating-areas
#> 
#> Access a single record by calling `bcdc_get_record(ID)` with the ID
#>  from the desired record.

# }