Select columns from bcdc_query_geodata() call

Similar to a dplyr::select call, this allows you to select which columns you want the Web Feature Service to return. A key difference between dplyr::select and bcdata::select is the presence of "sticky" columns that are returned regardless of what columns are selected. If any of these "sticky" columns are selected only "sticky" columns are return. bcdc_describe_feature is one way to tell if columns are sticky in advance of issuing the Web Feature Service call.

See dplyr::select for details.

Usage

# S3 method for bcdc_promise
select(.data, ...)

Arguments

.data

object of class bcdc_promise (likely passed from bcdc_query_geodata())

...

One or more unquoted expressions separated by commas. See details.

Methods (by class)

select(bcdc_promise): select.bcdc_promise

Examples

# \donttest{
try(
  feature_spec <- bcdc_describe_feature("bc-airports")
)

try(
  ## Columns that can selected:
  feature_spec[feature_spec$sticky == TRUE,]
)
#> # A tibble: 9 × 5
#>   col_name                    sticky remot…¹ local…² colum…³
#>   <chr>                       <lgl>  <chr>   <chr>   <chr>  
#> 1 id                          TRUE   xsd:st… charac… NA     
#> 2 CUSTODIAN_ORG_DESCRIPTION   TRUE   xsd:st… charac… CUSTOD…
#> 3 BUSINESS_CATEGORY_CLASS     TRUE   xsd:st… charac… BUSINE…
#> 4 BUSINESS_CATEGORY_DESCRIPT… TRUE   xsd:st… charac… BUSINE…
#> 5 OCCUPANT_TYPE_DESCRIPTION   TRUE   xsd:st… charac… OCCUPA…
#> 6 SOURCE_DATA_ID              TRUE   xsd:st… charac… SOURCE…
#> 7 SUPPLIED_SOURCE_ID_IND      TRUE   xsd:st… charac… SUPPLI…
#> 8 AIRPORT_NAME                TRUE   xsd:st… charac… AIRPOR…
#> 9 SEQUENCE_ID                 TRUE   xsd:de… numeric SEQUEN…
#> # … with abbreviated variable names ¹​remote_col_type,
#> #   ²​local_col_type, ³​column_comments

## Select columns
try(
  res <- bcdc_query_geodata("bc-airports") %>%
    select(DESCRIPTION, PHYSICAL_ADDRESS)
)

## Select "sticky" columns
try(
  res <- bcdc_query_geodata("bc-airports") %>%
    select(LOCALITY)
)
# }