Select columns from Web Service call

Similar to a dplyr::select call, this allows you to select which columns you want the Web Service to return. A key difference between dplyr::select and bcdata::select is the presence of "sticky" columns that are returned regardless of what columns are selected. If any of these "sticky" columns are selected only "sticky" columns are return. bcdc_describe_feature is one way to tell if columns are sticky in advance of issuing the Web Service call.

# S3 method for bcdc_promise
select(.data, ...)

Arguments

.data
object of class bcdc_promise (likely passed from bcdc_query_geodata())

...
One or more unquoted expressions separated by commas. See details.

Examples

# \donttest{
feature_spec <- bcdc_describe_feature("bc-airports")
## Columns that can selected:
feature_spec[feature_spec$nillable == TRUE,]

#> Warning: Unknown or uninitialised column: 'nillable'.

#> Warning: Length of logical index must be 1 or 42, not 0

#> # A tibble: 0 x 4
#> # … with 4 variables: col_name <chr>, selectable <lgl>, remote_col_type <chr>,
#> #   local_col_type <chr>

## Select columns
bcdc_query_geodata("bc-airports") %>%
  select(DESCRIPTION, PHYSICAL_ADDRESS)

#> Querying 'bc-airports' record
#> <U+25CF> Using collect() on this object will return 455 features and 14 fields
#> <U+25CF> At most six rows of the record are printed here
#> ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> Simple feature collection with 6 features and 14 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 833323.9 ymin: 381604.1 xmax: 1198292 ymax: 1054950
#> epsg (SRID):    3005
#> proj4string:    +proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 6 x 15
#>   id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>   <chr> <chr>            <chr>            <chr>            <chr>           
#> 1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 4 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 5 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 6 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 10 more variables: SOURCE_DATA_ID <chr>, SUPPLIED_SOURCE_ID_IND <chr>,
#> #   AIRPORT_NAME <chr>, DESCRIPTION <chr>, PHYSICAL_ADDRESS <chr>,
#> #   LOCALITY <chr>, LATITUDE <dbl>, LONGITUDE <dbl>, SEQUENCE_ID <int>,
#> #   geometry <POINT [m]>

## Select "sticky" columns
bcdc_query_geodata("bc-airports") %>%
  select(LOCALITY)

#> Querying 'bc-airports' record
#> <U+25CF> Using collect() on this object will return 455 features and 12 fields
#> <U+25CF> At most six rows of the record are printed here
#> ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> Simple feature collection with 6 features and 12 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 833323.9 ymin: 381604.1 xmax: 1198292 ymax: 1054950
#> epsg (SRID):    3005
#> proj4string:    +proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 6 x 13
#>   id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>   <chr> <chr>            <chr>            <chr>            <chr>           
#> 1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 4 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 5 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 6 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 8 more variables: SOURCE_DATA_ID <chr>, SUPPLIED_SOURCE_ID_IND <chr>,
#> #   AIRPORT_NAME <chr>, LOCALITY <chr>, LATITUDE <dbl>, LONGITUDE <dbl>,
#> #   SEQUENCE_ID <int>, geometry <POINT [m]>

# }