Describe a Web Service feature

Describe the columns from a Web Service feature. The column name, whether a column can be separated from the record in Web Service (nillable) and the type of column are returned. This can be a helpful tool to examine a layer before issuing a query with bcdc_query_geodata

bcdc_describe_feature(record)

Arguments

record
either a bcdc_record object (from the result of bcdc_get_record()) or a character string denoting the name or ID of a resource (or the URL).

It is advised to use the permament ID for a record rather than the human-readable name to guard against future name changes of the record. If you use the human-readable name a warning will be issued once per session. You can silence these warnings altogether by setting an option: options("silence_named_get_data_warning" = TRUE) - which you can set in your .Rprofile file so the option persists across sessions.

Examples

 bcdc_describe_feature("bc-airports")

#> Warning: It is advised to use the permanent id ('76b1b7a3-2112-4444-857a-afccf7b20da8') rather than the name of the record ('bc-airports') to guard against future name changes.

#> # A tibble: 42 x 4
#>    col_name                      selectable remote_col_type local_col_type
#>    <chr>                         <lgl>      <chr>           <chr>         
#>  1 id                            FALSE      xsd:string      character     
#>  2 CUSTODIAN_ORG_DESCRIPTION     FALSE      xsd:string      character     
#>  3 BUSINESS_CATEGORY_CLASS       FALSE      xsd:string      character     
#>  4 BUSINESS_CATEGORY_DESCRIPTION FALSE      xsd:string      character     
#>  5 OCCUPANT_TYPE_DESCRIPTION     FALSE      xsd:string      character     
#>  6 SOURCE_DATA_ID                FALSE      xsd:string      character     
#>  7 SUPPLIED_SOURCE_ID_IND        FALSE      xsd:string      character     
#>  8 AIRPORT_NAME                  FALSE      xsd:string      character     
#>  9 DESCRIPTION                   TRUE       xsd:string      character     
#> 10 PHYSICAL_ADDRESS              TRUE       xsd:string      character     
#> # … with 32 more rows