Download and read a resource from a B.C. Data Catalogue record

Download and read a resource from a B.C. Data Catalogue record

bcdc_get_data(record, resource = NULL, ...)

Arguments

record
either a bcdc_record object (from the result of bcdc_get_record()) or a character string denoting the name or ID of a resource (or the URL).

It is advised to use the permament ID for a record rather than the human-readable name to guard against future name changes of the record. If you use the human-readable name a warning will be issued once per session. You can silence these warnings altogether by setting an option: options("silence_named_get_data_warning" = TRUE) - which you can set in your .Rprofile file so the option persists across sessions.

resource
optional argument used when there are multiple data files within the same record. See examples.

...
arguments passed to other functions. Tabular data is passed to a function to handle the import based on the file extension. bcdc_read_functions() provides details on which functions handle the data import. You can then use this information to look at the help pages of those functions. See the examples for a workflow that illustrates this process. For spatial Web Service data the ... arguments are passed to bcdc_query_geodata().

Value

An object of a type relevant to the resource (usually a tibble or an sf object)

Examples

# NOT RUN {
# Using the record and resource ID:
bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8',
              resource = '4d0377d9-e8a1-429b-824f-0ce8f363512c')
bcdc_get_data('1d21922b-ec4f-42e5-8f6b-bf320a286157')

# Using a `bcdc_record` object obtained from `bcdc_get_record`:
record <- bcdc_get_record('1d21922b-ec4f-42e5-8f6b-bf320a286157')
bcdc_get_data(record)

## Example of correcting import problems

## Some initial problems reading in the data
bcdc_get_data('d7e6c8c7-052f-4f06-b178-74c02c243ea4')

## From bcdc_get_record we realize that the data is in xlsx format
bcdc_get_record('d7e6c8c7-052f-4f06-b178-74c02c243ea4')

## bcdc_read_functions let's us know that bcdata
## uses readxl::read_excel to import xlsx files
bcdc_read_functions()

## If you read the help page for readxl::read_excel,
## it seems likely that we need to skip the first row:
bcdc_get_data('d7e6c8c7-052f-4f06-b178-74c02c243ea4', skip = 1)

# }