Download and read a resource from a B.C. Data Catalogue record

Download and read a resource from a B.C. Data Catalogue record

bcdc_get_data(record, resource = NULL, verbose = TRUE, ...)

Arguments

record
either a bcdc_record object (from the result of bcdc_get_record()), a character string denoting the name or ID of a resource (or the URL) or a BC Geographic Warehouse (BCGW) name.

It is advised to use the permanent ID for a record or the BCGW name rather than the human-readable name to guard against future name changes of the record. If you use the human-readable name a warning will be issued once per session. You can silence these warnings altogether by setting an option: options("silence_named_get_data_warning" = TRUE) - which you can set in your .Rprofile file so the option persists across sessions.

resource
optional argument used when there are multiple data files within the same record. See examples.

verbose
When more than one resource is available for a record, should extra information about those resources be printed to the console? Default TRUE

...
arguments passed to other functions. Tabular data is passed to a function to handle the import based on the file extension. bcdc_read_functions() provides details on which functions handle the data import. You can then use this information to look at the help pages of those functions. See the examples for a workflow that illustrates this process. For spatial Web Service data the ... arguments are passed to bcdc_query_geodata().

Value

An object of a type relevant to the resource (usually a tibble or an sf object)

Examples

# \donttest{
# Using the record and resource ID:
bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8',
              resource = '4d0377d9-e8a1-429b-824f-0ce8f363512c')

#> Simple feature collection with 455 features and 41 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 406543.7 ymin: 367957.6 xmax: 1796645 ymax: 1689146
#> CRS:            3005
#> # A tibble: 455 x 42
#>    id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>  * <chr> <chr>            <chr>            <chr>            <chr>           
#>  1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  4 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  5 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  6 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  7 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  8 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  9 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 10 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 445 more rows, and 37 more variables: SOURCE_DATA_ID <chr>,
#> #   SUPPLIED_SOURCE_ID_IND <chr>, AIRPORT_NAME <chr>, DESCRIPTION <chr>,
#> #   PHYSICAL_ADDRESS <chr>, ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>,
#> #   POSTAL_CODE <chr>, LOCALITY <chr>, CONTACT_PHONE <chr>,
#> #   CONTACT_EMAIL <chr>, CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>,
#> #   LATITUDE <dbl>, LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <dbl>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

bcdc_get_data('1d21922b-ec4f-42e5-8f6b-bf320a286157')

#> Reading the data using the read_csv function from the readr package.

#> Parsed with column specification:
#> cols(
#>   CITY = col_character(),
#>   NAME = col_character(),
#>   ADDRESS = col_character(),
#>   POSTALCODE = col_character()
#> )

#> # A tibble: 357 x 4
#>    CITY        NAME                    ADDRESS                        POSTALCODE
#>    <chr>       <chr>                   <chr>                          <chr>     
#>  1 PRINCE GEO… Pacific Western Brewin… 641 NORTH NECHAKO RD           V2K4M4    
#>  2 OSOYOOS     Bordertown Vineyards    9140 92ND AVENUE               V0H1V2    
#>  3 CHILLIWACK  Whispering Horse Winery 43721 VEDDER MTN. ROAD         V2R4C5    
#>  4 KELOWNA     Okanaganvilla Estate W… 3240 POOLEY ROAD               V1W4G7    
#>  5 OLIVER      Church & State Wines (… 31120 87TH STREET              V0H1T0    
#>  6 KEREMEOS    Corcelettes Estate Win… 2582 UPPER BENCH ROAD          V0X1N4    
#>  7 DUNCAN      Blue Grouse Vineyards   4365 BLUE GROUSE ROAD AND 218… V9L6M3    
#>  8 VANCOUVER   City Side Winery        328 WEST 2ND AVENUE            V5Y1C8    
#>  9 WEST KELOW… Volcanic Hills Estate … 2845 BOUCHERIE ROAD            V1Z2G6    
#> 10 OLIVER      Orchard Hill Estate Ci… 3480 FRUITVALE WAY             V0H1T0    
#> # … with 347 more rows

# Using a `bcdc_record` object obtained from `bcdc_get_record`:
record <- bcdc_get_record('1d21922b-ec4f-42e5-8f6b-bf320a286157')
bcdc_get_data(record)

#> Reading the data using the read_csv function from the readr package.

#> Parsed with column specification:
#> cols(
#>   CITY = col_character(),
#>   NAME = col_character(),
#>   ADDRESS = col_character(),
#>   POSTALCODE = col_character()
#> )

#> # A tibble: 357 x 4
#>    CITY        NAME                    ADDRESS                        POSTALCODE
#>    <chr>       <chr>                   <chr>                          <chr>     
#>  1 PRINCE GEO… Pacific Western Brewin… 641 NORTH NECHAKO RD           V2K4M4    
#>  2 OSOYOOS     Bordertown Vineyards    9140 92ND AVENUE               V0H1V2    
#>  3 CHILLIWACK  Whispering Horse Winery 43721 VEDDER MTN. ROAD         V2R4C5    
#>  4 KELOWNA     Okanaganvilla Estate W… 3240 POOLEY ROAD               V1W4G7    
#>  5 OLIVER      Church & State Wines (… 31120 87TH STREET              V0H1T0    
#>  6 KEREMEOS    Corcelettes Estate Win… 2582 UPPER BENCH ROAD          V0X1N4    
#>  7 DUNCAN      Blue Grouse Vineyards   4365 BLUE GROUSE ROAD AND 218… V9L6M3    
#>  8 VANCOUVER   City Side Winery        328 WEST 2ND AVENUE            V5Y1C8    
#>  9 WEST KELOW… Volcanic Hills Estate … 2845 BOUCHERIE ROAD            V1Z2G6    
#> 10 OLIVER      Orchard Hill Estate Ci… 3480 FRUITVALE WAY             V0H1T0    
#> # … with 347 more rows

# Using a BCGW name
bcdc_get_data("WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW")

#> Simple feature collection with 455 features and 41 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 406543.7 ymin: 367957.6 xmax: 1796645 ymax: 1689146
#> CRS:            3005
#> # A tibble: 455 x 42
#>    id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>  * <chr> <chr>            <chr>            <chr>            <chr>           
#>  1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  4 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  5 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  6 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  7 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  8 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#>  9 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 10 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 445 more rows, and 37 more variables: SOURCE_DATA_ID <chr>,
#> #   SUPPLIED_SOURCE_ID_IND <chr>, AIRPORT_NAME <chr>, DESCRIPTION <chr>,
#> #   PHYSICAL_ADDRESS <chr>, ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>,
#> #   POSTAL_CODE <chr>, LOCALITY <chr>, CONTACT_PHONE <chr>,
#> #   CONTACT_EMAIL <chr>, CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>,
#> #   LATITUDE <dbl>, LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <dbl>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

## Example of correcting import problems

## Some initial problems reading in the data
bcdc_get_data('d7e6c8c7-052f-4f06-b178-74c02c243ea4')

#> Reading the data using the read_xlsx function from the readxl package.

#> 
#> This .xlsx resource contains the following sheets: 
#>  'Population'
#>  'Gender & Age Distribution'
#>  'Median, Average & Growth Rates'
#>  'Households'
#>  'Household Size'

#> Defaulting to the 'Population' sheet. See ?bcdc_get_data for examples on how to specify a sheet.

#> New names:
#> * `` -> ...2
#> * `` -> ...3
#> * `` -> ...4
#> * `` -> ...5
#> * `` -> ...6
#> * ...

#> # A tibble: 901 x 11
#>    `Regional Distri… ...2  ...3  ...4  ...5  ...6  ...7  ...8  ...9  ...10 ...11
#>    <chr>             <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 RD Code           Regi… Sex   Year  Age … 15-19 20-24 25-64 65-84 85+   All …
#>  2 1000              East… Male  2019  4863  1651  1838  18148 5793  543   32836
#>  3 1000              East… Male  2020  4850  1607  1812  18083 6022  562   32936
#>  4 1000              East… Male  2021  4850  1569  1819  17918 6295  592   33043
#>  5 1000              East… Male  2022  4785  1594  1693  17942 6533  594   33141
#>  6 1000              East… Male  2023  4756  1577  1622  17910 6794  579   33238
#>  7 1000              East… Male  2024  4731  1574  1561  17848 7023  593   33330
#>  8 1000              East… Male  2025  4673  1604  1512  17730 7226  670   33415
#>  9 1000              East… Male  2026  4590  1659  1483  17589 7495  668   33484
#> 10 1000              East… Male  2027  4515  1645  1510  17430 7705  731   33536
#> # … with 891 more rows

## From bcdc_get_record we realize that the data is in xlsx format
bcdc_get_record('8620ce82-4943-43c4-9932-40730a0255d6')

#> B.C. Data Catalogue Record: New Homes Registry (2016-2018) 
#> 
#> Name: new-homes-registry-2016-2018- (ID: 8620ce82-4943-43c4-9932-40730a0255d6 )
#> Permalink: https://catalogue.data.gov.bc.ca/dataset/8620ce82-4943-43c4-9932-40730a0255d6
#> Sector: Economy
#> Licence: Access Only
#> Type: Dataset
#> Last Updated: 2020-04-23 
#> 
#> Description: This dataset contains information from BC Housing's New Homes Registry.  
#> 
#> Resources: (1)
#> # A tibble: 1 x 8
#>   name    url        id       format ext   package_id  location bcdata_available
#>   <chr>   <chr>      <chr>    <chr>  <chr> <chr>       <chr>    <lgl>           
#> 1 New Ho… https://c… 32ff8d7… xlsx   xlsx  8620ce82-4… catalog… TRUE            
#> 
#> You can access the 'Resources' data frame using bcdc_tidy_resources()
#> 
#> 

## bcdc_read_functions let's us know that bcdata
## uses readxl::read_excel to import xlsx files
bcdc_read_functions()

#> # A tibble: 11 x 3
#>    format  package fun      
#>    <chr>   <chr>   <chr>    
#>  1 kml     sf      read_sf  
#>  2 geojson sf      read_sf  
#>  3 gpkg    sf      read_sf  
#>  4 gdb     sf      read_sf  
#>  5 fgdb    sf      read_sf  
#>  6 shp     sf      read_sf  
#>  7 csv     readr   read_csv 
#>  8 txt     readr   read_tsv 
#>  9 tsv     readr   read_tsv 
#> 10 xlsx    readxl  read_xlsx
#> 11 xls     readxl  read_xls 

## bcdata let's you know that this resource has
## multiple worksheets
bcdc_get_data('8620ce82-4943-43c4-9932-40730a0255d6')

#> Reading the data using the read_xlsx function from the readxl package.

#> 
#> This .xlsx resource contains the following sheets: 
#>  'Notes'
#>  'Regional Districts'
#>  'Single Detached'
#>  'Multi Unit Homes'
#>  'Purpose Built Rental'

#> Defaulting to the 'Notes' sheet. See ?bcdc_get_data for examples on how to specify a sheet.

#> # A tibble: 0 x 0

## we can control what is read in from an excel file
## using arguments from readxl::read_excel

bcdc_get_data('8620ce82-4943-43c4-9932-40730a0255d6', sheet = 'Regional Districts')

#> Reading the data using the read_xlsx function from the readxl package.

#> 
#> This .xlsx resource contains the following sheets: 
#>  'Notes'
#>  'Regional Districts'
#>  'Single Detached'
#>  'Multi Unit Homes'
#>  'Purpose Built Rental'

#> New names:
#> * `` -> ...2
#> * `` -> ...3
#> * `` -> ...4
#> * `` -> ...5
#> * `` -> ...6
#> * ...

#> # A tibble: 31 x 10
#>    `Registered New Hom… ...2     ...3  ...4 ...5    ...6  ...7 ...8   ...9 ...10
#>    <chr>                <chr>   <dbl> <dbl> <chr>  <dbl> <dbl> <chr> <dbl> <dbl>
#>  1 Regional District    Single…    NA    NA Multi…    NA    NA Rent…    NA    NA
#>  2 NA                   2016     2017  2018 2016    2017  2018 2016   2017  2018
#>  3 Alberni-Clayoquot    98        105   111 32        17    11 24        0    35
#>  4 Bulkley-Nechako      63         51    42 17        10     0 0        19     0
#>  5 Capital              1123      982   947 1186    1359  2085 257    1658  1184
#>  6 Cariboo              113       105   154 8          6     0 0        70    41
#>  7 Central Coast        0           1     1 0          0     0 0         0     0
#>  8 Central Kootenay     147       140   171 21        30    18 3        22    15
#>  9 Central Okanagan     853       889   675 1145     823  1499 790    1165   534
#> 10 Columbia-Shuswap     227       285   291 18        90   143 0         0     3
#> # … with 21 more rows

# }