Show a single B.C. Data Catalogue record

Show a single B.C. Data Catalogue record

bcdc_get_record(id)

Arguments

id
the human-readable name, permalink ID, or URL of the record.

It is advised to use the permanent ID for a record rather than the human-readable name to guard against future name changes of the record. If you use the human-readable name a warning will be issued once per session. You can silence these warnings altogether by setting an option: options("silence_named_get_record_warning" = TRUE) - which you can put in your .Rprofile file so the option persists across sessions.

Value

A list containing the metadata for the record

Examples

# \donttest{
bcdc_get_record("https://catalogue.data.gov.bc.ca/dataset/bc-airports")

#> B.C. Data Catalogue Record:
#>     BC Airports 
#> 
#> Name: bc-airports (ID: 76b1b7a3-2112-4444-857a-afccf7b20da8 )
#> Permalink: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8
#> Sector: Natural Resources
#> Licence: Open Government Licence - British Columbia
#> Type: Geographic
#> Last Updated: 2019-09-05 
#> 
#> Description:
#>     BC Airports identifies locations where aircraft may take-off and land. No guarantee
#>     is given that an identified point will be maintained to sufficient standards for
#>     landing and take-off of any/all aircraft.  It includes airports, aerodromes, water
#>     aerodromes, heliports, and airstrips. 
#> 
#> Resources: ( 4 )
#> 1) BC Geographic Warehouse Custom Download
#>      format: other 
#>      url: https://catalogue.data.gov.bc.ca/api/ofi/other/WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW 
#>      resource: 604c8be1-b3f3-45b4-8030-5f9c8be71645 
#>      available in R via bcdata:  FALSE 
#> 
#> 2) BC_Airports_Attribute_Values
#>      format: xlsx 
#>      url: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8/resource/fcccba36-b528-4731-8978-940b3cc04f69/download/wilmbvicgeobccriticalinfrastructuredocumentsbcairportsattributevalues.xlsx 
#>      resource: fcccba36-b528-4731-8978-940b3cc04f69 
#>      available in R via bcdata:  TRUE 
#>      code: bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8', resource = 'fcccba36-b528-4731-8978-940b3cc04f69')
#> 
#> 3) WFS request (Spatial Data)
#>      format: wfs 
#>      resource: 4d0377d9-e8a1-429b-824f-0ce8f363512c 
#>      available in R via bcdata:  TRUE 
#>      code: bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8', resource = '4d0377d9-e8a1-429b-824f-0ce8f363512c')
#> 
#> 4) KML Network Link
#>      format: kml 
#>      url: http://openmaps.gov.bc.ca/kml/geo/layers/WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW_loader.kml 
#>      resource: 5b9f22bf-eaa6-4640-ae65-9d3600672745 
#>      available in R via bcdata:  FALSE 
#> 

bcdc_get_record("bc-airports")

#> B.C. Data Catalogue Record:
#>     BC Airports 
#> 
#> Name: bc-airports (ID: 76b1b7a3-2112-4444-857a-afccf7b20da8 )
#> Permalink: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8
#> Sector: Natural Resources
#> Licence: Open Government Licence - British Columbia
#> Type: Geographic
#> Last Updated: 2019-09-05 
#> 
#> Description:
#>     BC Airports identifies locations where aircraft may take-off and land. No guarantee
#>     is given that an identified point will be maintained to sufficient standards for
#>     landing and take-off of any/all aircraft.  It includes airports, aerodromes, water
#>     aerodromes, heliports, and airstrips. 
#> 
#> Resources: ( 4 )
#> 1) BC Geographic Warehouse Custom Download
#>      format: other 
#>      url: https://catalogue.data.gov.bc.ca/api/ofi/other/WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW 
#>      resource: 604c8be1-b3f3-45b4-8030-5f9c8be71645 
#>      available in R via bcdata:  FALSE 
#> 
#> 2) BC_Airports_Attribute_Values
#>      format: xlsx 
#>      url: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8/resource/fcccba36-b528-4731-8978-940b3cc04f69/download/wilmbvicgeobccriticalinfrastructuredocumentsbcairportsattributevalues.xlsx 
#>      resource: fcccba36-b528-4731-8978-940b3cc04f69 
#>      available in R via bcdata:  TRUE 
#>      code: bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8', resource = 'fcccba36-b528-4731-8978-940b3cc04f69')
#> 
#> 3) WFS request (Spatial Data)
#>      format: wfs 
#>      resource: 4d0377d9-e8a1-429b-824f-0ce8f363512c 
#>      available in R via bcdata:  TRUE 
#>      code: bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8', resource = '4d0377d9-e8a1-429b-824f-0ce8f363512c')
#> 
#> 4) KML Network Link
#>      format: kml 
#>      url: http://openmaps.gov.bc.ca/kml/geo/layers/WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW_loader.kml 
#>      resource: 5b9f22bf-eaa6-4640-ae65-9d3600672745 
#>      available in R via bcdata:  FALSE 
#> 

bcdc_get_record("https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8")

#> B.C. Data Catalogue Record:
#>     BC Airports 
#> 
#> Name: bc-airports (ID: 76b1b7a3-2112-4444-857a-afccf7b20da8 )
#> Permalink: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8
#> Sector: Natural Resources
#> Licence: Open Government Licence - British Columbia
#> Type: Geographic
#> Last Updated: 2019-09-05 
#> 
#> Description:
#>     BC Airports identifies locations where aircraft may take-off and land. No guarantee
#>     is given that an identified point will be maintained to sufficient standards for
#>     landing and take-off of any/all aircraft.  It includes airports, aerodromes, water
#>     aerodromes, heliports, and airstrips. 
#> 
#> Resources: ( 4 )
#> 1) BC Geographic Warehouse Custom Download
#>      format: other 
#>      url: https://catalogue.data.gov.bc.ca/api/ofi/other/WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW 
#>      resource: 604c8be1-b3f3-45b4-8030-5f9c8be71645 
#>      available in R via bcdata:  FALSE 
#> 
#> 2) BC_Airports_Attribute_Values
#>      format: xlsx 
#>      url: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8/resource/fcccba36-b528-4731-8978-940b3cc04f69/download/wilmbvicgeobccriticalinfrastructuredocumentsbcairportsattributevalues.xlsx 
#>      resource: fcccba36-b528-4731-8978-940b3cc04f69 
#>      available in R via bcdata:  TRUE 
#>      code: bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8', resource = 'fcccba36-b528-4731-8978-940b3cc04f69')
#> 
#> 3) WFS request (Spatial Data)
#>      format: wfs 
#>      resource: 4d0377d9-e8a1-429b-824f-0ce8f363512c 
#>      available in R via bcdata:  TRUE 
#>      code: bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8', resource = '4d0377d9-e8a1-429b-824f-0ce8f363512c')
#> 
#> 4) KML Network Link
#>      format: kml 
#>      url: http://openmaps.gov.bc.ca/kml/geo/layers/WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW_loader.kml 
#>      resource: 5b9f22bf-eaa6-4640-ae65-9d3600672745 
#>      available in R via bcdata:  FALSE 
#> 

bcdc_get_record("76b1b7a3-2112-4444-857a-afccf7b20da8")

#> B.C. Data Catalogue Record:
#>     BC Airports 
#> 
#> Name: bc-airports (ID: 76b1b7a3-2112-4444-857a-afccf7b20da8 )
#> Permalink: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8
#> Sector: Natural Resources
#> Licence: Open Government Licence - British Columbia
#> Type: Geographic
#> Last Updated: 2019-09-05 
#> 
#> Description:
#>     BC Airports identifies locations where aircraft may take-off and land. No guarantee
#>     is given that an identified point will be maintained to sufficient standards for
#>     landing and take-off of any/all aircraft.  It includes airports, aerodromes, water
#>     aerodromes, heliports, and airstrips. 
#> 
#> Resources: ( 4 )
#> 1) BC Geographic Warehouse Custom Download
#>      format: other 
#>      url: https://catalogue.data.gov.bc.ca/api/ofi/other/WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW 
#>      resource: 604c8be1-b3f3-45b4-8030-5f9c8be71645 
#>      available in R via bcdata:  FALSE 
#> 
#> 2) BC_Airports_Attribute_Values
#>      format: xlsx 
#>      url: https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8/resource/fcccba36-b528-4731-8978-940b3cc04f69/download/wilmbvicgeobccriticalinfrastructuredocumentsbcairportsattributevalues.xlsx 
#>      resource: fcccba36-b528-4731-8978-940b3cc04f69 
#>      available in R via bcdata:  TRUE 
#>      code: bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8', resource = 'fcccba36-b528-4731-8978-940b3cc04f69')
#> 
#> 3) WFS request (Spatial Data)
#>      format: wfs 
#>      resource: 4d0377d9-e8a1-429b-824f-0ce8f363512c 
#>      available in R via bcdata:  TRUE 
#>      code: bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8', resource = '4d0377d9-e8a1-429b-824f-0ce8f363512c')
#> 
#> 4) KML Network Link
#>      format: kml 
#>      url: http://openmaps.gov.bc.ca/kml/geo/layers/WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW_loader.kml 
#>      resource: 5b9f22bf-eaa6-4640-ae65-9d3600672745 
#>      available in R via bcdata:  FALSE 
#> 

# }