Query data from the B.C. Web Service

Queries features from the B.C. Web Service. The data must be available as a Web Service. See bcdc_get_record(record)$resources). If the record is greater than 10000 rows, the response will be paginated. If you are querying layers of this size, expect that the request will take quite a while.

bcdc_query_geodata(record, crs = 3005)

Arguments

record
either a bcdc_record object (from the result of bcdc_get_record()), a character string denoting the name or ID of a resource (or the URL) or a BC Geographic Warehouse (BCGW) name.

It is advised to use the permanent ID for a record or the BCGW name rather than the human-readable name to guard against future name changes of the record. If you use the human-readable name a warning will be issued once per session. You can silence these warnings altogether by setting an option: options("silence_named_get_data_warning" = TRUE) - which you can set in your .Rprofile file so the option persists across sessions.

crs
the epsg code for the coordinate reference system. Defaults to 3005 (B.C. Albers). See https://epsg.io.

Value

A bcdc_promise object. This object includes all of the information required to retrieve the requested data. In order to get the actual data as an sf object, you need to run collect() on the bcdc_promise.

Details

Note that this function doesn't actually return the data, but rather an object of class bcdc_promise, which includes all of the information required to retrieve the requested data. In order to get the actual data as an sf object, you need to run collect() on the bcdc_promise. This allows further refining the call to bcdc_query_geodata() with filter() and/or select() statements before pulling down the actual data as an sf object with collect(). See examples.

Examples

# \donttest{
# Returns a bcdc_promise, which can be further refined using filter/select:
try(
  bcdc_query_geodata("bc-airports", crs = 3857)
)

#> Querying 'bc-airports' record
#> ● Using collect() on this object will return 455
#> ● features and 41 fields
#> ● At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 41 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -14313270 ymin: 6176688 xmax: -13728510 ymax: 7259410
#> projected CRS:  WGS 84 / Pseudo-Mercator
#> # A tibble: 6 x 42
#>   id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>   <chr> <chr>            <chr>            <chr>            <chr>           
#> 1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 4 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 5 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 6 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 37 more variables: SOURCE_DATA_ID <chr>, SUPPLIED_SOURCE_ID_IND <chr>,
#> #   AIRPORT_NAME <chr>, DESCRIPTION <chr>, PHYSICAL_ADDRESS <chr>,
#> #   ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>, POSTAL_CODE <chr>,
#> #   LOCALITY <chr>, CONTACT_PHONE <chr>, CONTACT_EMAIL <chr>,
#> #   CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>, LATITUDE <dbl>,
#> #   LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <dbl>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

# To obtain the actual data as an sf object, collect() must be called:
try(
  bcdc_query_geodata("bc-airports", crs = 3857) %>%
    filter(PHYSICAL_ADDRESS == 'Victoria, BC') %>%
    collect()
)

#> Error : There was an issue sending this WFS request
#> ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄
#> Request:
#>   URL: https://openmaps.gov.bc.ca/geo/pub/wfs/
#>   POST fields:
#>     SERVICE=WFS&VERSION=2.0.0&REQUEST=GetFeature&outputFormat=application%2Fjson&typeNames=WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW&SRSNAME=EPSG%3A3857&CQL_FILTER=%28%22PHYSICAL_ADDRESS%22%20%3D%20%27Victoria%2C%20BC%27%29
#>   Content-Type: application/x-www-form-urlencoded
#>   Accept-Encoding: gzip, deflate
#>   Accept: application/json, text/xml, application/xml, */*
#>   User-Agent: https://github.com/bcgov/bcdata
#> Response:
#>   status: HTTP/1.1 400 
#>   date: Thu, 22 Oct 2020 20:34:09 GMT
#>   server: Apache
#>   x-frame-options: allow-from (null)
#>   x-control-flow-delay-ms: 0
#>   content-type: application/xml
#>   set-cookie: GS_FLOW_CONTROL=GS_CFLOW_6d23bbc1:1755204ad8d:-7fb9
#>   access-control-allow-origin: (null)
#>   access-control-allow-credentials: true
#>   access-control-allow-methods: POST, GET, OPTIONS, HEAD
#>   access-control-allow-headers: X-Requested-With, Referer, Origin, Content-Type, SOAPAction, Authorization, Accept
#>   access-control-max-age: 1000
#>   connection: close
#>   transfer-encoding: chunked
#> 

try(
  bcdc_query_geodata("groundwater-wells") %>%
    filter(OBSERVATION_WELL_NUMBER == "108") %>%
    select(WELL_TAG_NUMBER, INTENDED_WATER_USE) %>%
    collect()
)

#> Error : There was an issue sending this WFS request
#> ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄
#> Request:
#>   URL: https://openmaps.gov.bc.ca/geo/pub/wfs/
#>   POST fields:
#>     resultType=hits&SERVICE=WFS&VERSION=2.0.0&REQUEST=GetFeature&outputFormat=application%2Fjson&typeNames=WHSE_WATER_MANAGEMENT.GW_WATER_WELLS_WRBC_SVW&SRSNAME=EPSG%3A3005&CQL_FILTER=%28%22OBSERVATION_WELL_NUMBER%22%20%3D%20%27108%27%29
#>   Content-Type: application/x-www-form-urlencoded
#>   Accept-Encoding: gzip, deflate
#>   Accept: application/json, text/xml, application/xml, */*
#>   User-Agent: https://github.com/bcgov/bcdata
#> Response:
#>   status: HTTP/1.1 400 
#>   date: Thu, 22 Oct 2020 20:34:21 GMT
#>   server: Apache
#>   x-frame-options: allow-from (null)
#>   x-control-flow-delay-ms: 0
#>   content-type: application/xml
#>   set-cookie: GS_FLOW_CONTROL=GS_CFLOW_-60e76474:17551ed38fb:7fba
#>   access-control-allow-origin: (null)
#>   access-control-allow-credentials: true
#>   access-control-allow-methods: POST, GET, OPTIONS, HEAD
#>   access-control-allow-headers: X-Requested-With, Referer, Origin, Content-Type, SOAPAction, Authorization, Accept
#>   access-control-max-age: 1000
#>   connection: close
#>   transfer-encoding: chunked
#> 

## A moderately large layer
try(
  bcdc_query_geodata("bc-environmental-monitoring-locations")
)

#> Error: There was an issue sending this WFS request
#> ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄
#> Request:
#>   URL: https://openmaps.gov.bc.ca/geo/pub/wfs/
#>   POST fields:
#>     resultType=hits&SERVICE=WFS&VERSION=2.0.0&REQUEST=GetFeature&outputFormat=application%2Fjson&typeNames=WHSE_ENVIRONMENTAL_MONITORING.EMS_MONITORING_LOCN_TYPES_SVW&SRSNAME=EPSG%3A3005
#>   Content-Type: application/x-www-form-urlencoded
#>   Accept-Encoding: gzip, deflate
#>   Accept: application/json, text/xml, application/xml, */*
#>   User-Agent: https://github.com/bcgov/bcdata
#> Response:
#>   status: HTTP/1.1 400 
#>   date: Thu, 22 Oct 2020 20:34:33 GMT
#>   server: Apache
#>   x-frame-options: allow-from (null)
#>   x-control-flow-delay-ms: 0
#>   content-type: application/xml
#>   set-cookie: GS_FLOW_CONTROL=GS_CFLOW_1e0b0ec5:1755201a5ae:2389
#>   access-control-allow-origin: (null)
#>   access-control-allow-credentials: true
#>   access-control-allow-methods: POST, GET, OPTIONS, HEAD
#>   access-control-allow-headers: X-Requested-With, Referer, Origin, Content-Type, SOAPAction, Authorization, Accept
#>   access-control-max-age: 1000
#>   connection: close
#>   transfer-encoding: chunked

try(
  bcdc_query_geodata("bc-environmental-monitoring-locations") %>%
    filter(PERMIT_RELATIONSHIP == "DISCHARGE")
)

#> Error: There was an issue sending this WFS request
#> ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄
#> Request:
#>   URL: https://openmaps.gov.bc.ca/geo/pub/wfs/
#>   POST fields:
#>     SERVICE=WFS&VERSION=2.0.0&REQUEST=GetFeature&outputFormat=application%2Fjson&typeNames=WHSE_ENVIRONMENTAL_MONITORING.EMS_MONITORING_LOCN_TYPES_SVW&SRSNAME=EPSG%3A3005&CQL_FILTER=%28%22PERMIT_RELATIONSHIP%22%20%3D%20%27DISCHARGE%27%29&count=6
#>   Content-Type: application/x-www-form-urlencoded
#>   Accept-Encoding: gzip, deflate
#>   Accept: application/json, text/xml, application/xml, */*
#>   User-Agent: https://github.com/bcgov/bcdata
#> Response:
#>   status: HTTP/1.1 400 
#>   date: Thu, 22 Oct 2020 20:34:43 GMT
#>   server: Apache
#>   x-frame-options: allow-from (null)
#>   x-control-flow-delay-ms: 0
#>   content-type: application/xml
#>   set-cookie: GS_FLOW_CONTROL=GS_CFLOW_-60e76474:1755204dd7c:-7f44
#>   access-control-allow-origin: (null)
#>   access-control-allow-credentials: true
#>   access-control-allow-methods: POST, GET, OPTIONS, HEAD
#>   access-control-allow-headers: X-Requested-With, Referer, Origin, Content-Type, SOAPAction, Authorization, Accept
#>   access-control-max-age: 1000
#>   connection: close
#>   transfer-encoding: chunked

## A very large layer
try(
  bcdc_query_geodata("terrestrial-protected-areas-representation-by-biogeoclimatic-unit")
)

#> Error: There was an issue sending this WFS request
#> ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄
#> Request:
#>   URL: https://openmaps.gov.bc.ca/geo/pub/wfs/
#>   POST fields:
#>     SERVICE=WFS&VERSION=2.0.0&REQUEST=GetFeature&outputFormat=application%2Fjson&typeNames=WHSE_LAND_AND_NATURAL_RESOURCE.PASO_PCT_BGC_CODE_PRTCTD_SVW&SRSNAME=EPSG%3A3005&count=6
#>   Content-Type: application/x-www-form-urlencoded
#>   Accept-Encoding: gzip, deflate
#>   Accept: application/json, text/xml, application/xml, */*
#>   User-Agent: https://github.com/bcgov/bcdata
#> Response:
#>   status: HTTP/1.1 400 
#>   date: Thu, 22 Oct 2020 20:34:55 GMT
#>   server: Apache
#>   x-frame-options: allow-from (null)
#>   x-control-flow-delay-ms: 0
#>   content-type: application/xml
#>   set-cookie: GS_FLOW_CONTROL=GS_CFLOW_-60e76474:1755204dd7c:-7f05
#>   access-control-allow-origin: (null)
#>   access-control-allow-credentials: true
#>   access-control-allow-methods: POST, GET, OPTIONS, HEAD
#>   access-control-allow-headers: X-Requested-With, Referer, Origin, Content-Type, SOAPAction, Authorization, Accept
#>   access-control-max-age: 1000
#>   connection: close
#>   transfer-encoding: chunked

## Using a BCGW name
try(
  bcdc_query_geodata("WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW")
)

#> Querying 'WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW' record
#> ● Using collect() on this object will return 455
#> ● features and 41 fields
#> ● At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 41 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 833323.9 ymin: 381604.1 xmax: 1198292 ymax: 1054950
#> projected CRS:  NAD83 / BC Albers
#> # A tibble: 6 x 42
#>   id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>   <chr> <chr>            <chr>            <chr>            <chr>           
#> 1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 4 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 5 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 6 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 37 more variables: SOURCE_DATA_ID <chr>, SUPPLIED_SOURCE_ID_IND <chr>,
#> #   AIRPORT_NAME <chr>, DESCRIPTION <chr>, PHYSICAL_ADDRESS <chr>,
#> #   ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>, POSTAL_CODE <chr>,
#> #   LOCALITY <chr>, CONTACT_PHONE <chr>, CONTACT_EMAIL <chr>,
#> #   CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>, LATITUDE <dbl>,
#> #   LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <dbl>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

# }