Query data from the B.C. Web Service

Queries features from the B.C. Web Service. The data must be available as a Web Service. See bcdc_get_record(record)$resources). If the record is greater than 10000 rows, the response will be paginated. If you are querying layers of this size, expect that the request will take quite a while.

bcdc_query_geodata(record, crs = 3005)

Arguments

record
either a bcdc_record object (from the result of bcdc_get_record()), a character string denoting the name or ID of a resource (or the URL) or a BC Geographic Warehouse (BCGW) name.

It is advised to use the permanent ID for a record or the BCGW name rather than the human-readable name to guard against future name changes of the record. If you use the human-readable name a warning will be issued once per session. You can silence these warnings altogether by setting an option: options("silence_named_get_data_warning" = TRUE) - which you can set in your .Rprofile file so the option persists across sessions.

crs
the epsg code for the coordinate reference system. Defaults to 3005 (B.C. Albers). See https://epsg.io.

Value

A bcdc_promise object. This object includes all of the information required to retrieve the requested data. In order to get the actual data as an sf object, you need to run collect() on the bcdc_promise.

Details

Note that this function doesn't actually return the data, but rather an object of class bcdc_promise, which includes all of the information required to retrieve the requested data. In order to get the actual data as an sf object, you need to run collect() on the bcdc_promise. This allows further refining the call to bcdc_query_geodata() with filter() and/or select() statements before pulling down the actual data as an sf object with collect(). See examples.

Examples

# \donttest{
# Returns a bcdc_promise, which can be further refined using filter/select:
bcdc_query_geodata("bc-airports", crs = 3857)

#> Querying 'bc-airports' record
#> ● Using collect() on this object will return 455 features and 41 fields
#> ● At most six rows of the record are printed here
#> ─────────────────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 41 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -14313270 ymin: 6176688 xmax: -13728510 ymax: 7259410
#> epsg (SRID):    3857
#> proj4string:    +proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +wktext +no_defs
#> # A tibble: 6 x 42
#>   id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>   <chr> <chr>            <chr>            <chr>            <chr>           
#> 1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 4 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 5 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 6 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 37 more variables: SOURCE_DATA_ID <chr>, SUPPLIED_SOURCE_ID_IND <chr>,
#> #   AIRPORT_NAME <chr>, DESCRIPTION <chr>, PHYSICAL_ADDRESS <chr>,
#> #   ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>, POSTAL_CODE <chr>,
#> #   LOCALITY <chr>, CONTACT_PHONE <chr>, CONTACT_EMAIL <chr>,
#> #   CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>, LATITUDE <dbl>,
#> #   LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <dbl>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

# To obtain the actual data as an sf object, collect() must be called:
bcdc_query_geodata("bc-airports", crs = 3857) %>%
  filter(PHYSICAL_ADDRESS == 'Victoria, BC') %>%
  collect()

#> Simple feature collection with 3 features and 41 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -13735460 ymin: 6176688 xmax: -13733670 ymax: 6177610
#> epsg (SRID):    3857
#> proj4string:    +proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +wktext +no_defs
#> # A tibble: 3 x 42
#>   id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#> * <chr> <chr>            <chr>            <chr>            <chr>           
#> 1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 37 more variables: SOURCE_DATA_ID <chr>, SUPPLIED_SOURCE_ID_IND <chr>,
#> #   AIRPORT_NAME <chr>, DESCRIPTION <chr>, PHYSICAL_ADDRESS <chr>,
#> #   ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>, POSTAL_CODE <chr>,
#> #   LOCALITY <chr>, CONTACT_PHONE <chr>, CONTACT_EMAIL <chr>,
#> #   CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>, LATITUDE <dbl>,
#> #   LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <int>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

bcdc_query_geodata("ground-water-wells") %>%
  filter(OBSERVATION_WELL_NUMBER == 108) %>%
  select(WELL_TAG_NUMBER, WATERSHED_CODE) %>%
  collect()

#> Simple feature collection with 1 feature and 9 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1273428 ymin: 1197388 xmax: 1273428 ymax: 1197388
#> epsg (SRID):    3005
#> proj4string:    +proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 1 x 10
#>   id    WELL_TAG_NO OBJECTID WELL_ID WELL_TAG_NUMBER WATERSHED_CODE WHEN_CREATED
#> * <chr> <chr>          <int>   <int>           <int> <chr>          <date>      
#> 1 WHSE… 0000000227… 59250828   57864           22722 0              2003-08-13  
#> # … with 3 more variables: WHO_CREATED <chr>,
#> #   WELL_LICENCE_GENERAL_STATUS <chr>, geometry <POINT [m]>

## A moderately large layer
bcdc_query_geodata("bc-environmental-monitoring-locations")

#> Querying 'bc-environmental-monitoring-locations' record
#> ● Using collect() on this object will return 37048 features and 35 fields
#> ● At most six rows of the record are printed here
#> ─────────────────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 35 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1424767 ymin: 465316.9 xmax: 1475694 ymax: 602912.2
#> epsg (SRID):    3005
#> proj4string:    +proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 6 x 36
#>   id    MONITORING_LOCA… CLIENT_LOCATION… CONTAMINATED_SI… DISCHARGE_MEDIA…
#>   <chr> <chr>                       <int> <chr>            <chr>           
#> 1 WHSE… 0500892                        52 NA               NA              
#> 2 WHSE… 0500893                        52 NA               NA              
#> 3 WHSE… 0500894                        52 NA               NA              
#> 4 WHSE… 0500895                        52 NA               NA              
#> 5 WHSE… 0500897                        52 NA               NA              
#> 6 WHSE… 0500898                        52 NA               NA              
#> # … with 31 more variables: ECO_REGION_CD <chr>, GEOREFERENCE_SOURCE_CD <chr>,
#> #   LOCATION_PURPOSE_CD <chr>, LOCATION_TYPE_CD <chr>, OFFICE_CD <chr>,
#> #   SOURCE_CLASS_CD <chr>, WATERSHED_CD <chr>, CALC_STATISTICS_IND <chr>,
#> #   LOCATION_AREA_IND <chr>, MOBILE_LOCATION_IND <chr>, CLOSED_DATE <chr>,
#> #   DESCRIPTION <chr>, ELEVATION <chr>, ESTABLISHED_DATE <date>,
#> #   LATITUDE <dbl>, LONGITUDE <dbl>, LEGAL_DESCRIPTION <chr>,
#> #   MONITORING_LOCN_COMMENT <chr>, MONITORING_LOCATION_NAME <chr>,
#> #   SOURCE_CLASS_NARRATIVE <chr>, WELL_IDENTIFICATION_PLATE_NO <chr>,
#> #   FIRST_SAMPLE_DATE <date>, LAST_SAMPLE_DATE <date>, SAMPLE_COUNT <int>,
#> #   RESULT_COUNT <int>, PERMIT_ID <int>, PERMIT_RELATIONSHIP <chr>,
#> #   FEATURE_CODE <chr>, OBJECTID <int>, SE_ANNO_CAD_DATA <chr>, geometry <POINT
#> #   [m]>

bcdc_query_geodata("bc-environmental-monitoring-locations") %>%
  filter(PERMIT_RELATIONSHIP == "DISCHARGE")

#> Querying 'bc-environmental-monitoring-locations' record
#> ● Using collect() on this object will return 11786 features and 35 fields
#> ● At most six rows of the record are printed here
#> ─────────────────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 35 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1249115 ymin: 474723.7 xmax: 1511701 ymax: 896127.6
#> epsg (SRID):    3005
#> proj4string:    +proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 6 x 36
#>   id    MONITORING_LOCA… CLIENT_LOCATION… CONTAMINATED_SI… DISCHARGE_MEDIA…
#>   <chr> <chr>                       <int> <chr>            <chr>           
#> 1 WHSE… E105573                        41 NA               4               
#> 2 WHSE… E105574                        52 NA               4               
#> 3 WHSE… E105575                        52 NA               4               
#> 4 WHSE… 0600174                        48 NA               1               
#> 5 WHSE… E105586                        52 NA               1               
#> 6 WHSE… E105587                        41 NA               3               
#> # … with 31 more variables: ECO_REGION_CD <chr>, GEOREFERENCE_SOURCE_CD <chr>,
#> #   LOCATION_PURPOSE_CD <chr>, LOCATION_TYPE_CD <chr>, OFFICE_CD <chr>,
#> #   SOURCE_CLASS_CD <chr>, WATERSHED_CD <chr>, CALC_STATISTICS_IND <chr>,
#> #   LOCATION_AREA_IND <chr>, MOBILE_LOCATION_IND <chr>, CLOSED_DATE <chr>,
#> #   DESCRIPTION <chr>, ELEVATION <chr>, ESTABLISHED_DATE <date>,
#> #   LATITUDE <dbl>, LONGITUDE <dbl>, LEGAL_DESCRIPTION <chr>,
#> #   MONITORING_LOCN_COMMENT <chr>, MONITORING_LOCATION_NAME <chr>,
#> #   SOURCE_CLASS_NARRATIVE <chr>, WELL_IDENTIFICATION_PLATE_NO <chr>,
#> #   FIRST_SAMPLE_DATE <date>, LAST_SAMPLE_DATE <date>, SAMPLE_COUNT <int>,
#> #   RESULT_COUNT <int>, PERMIT_ID <int>, PERMIT_RELATIONSHIP <chr>,
#> #   FEATURE_CODE <chr>, OBJECTID <int>, SE_ANNO_CAD_DATA <chr>, geometry <POINT
#> #   [m]>

## A very large layer
bcdc_query_geodata("terrestrial-protected-areas-representation-by-biogeoclimatic-unit")

#> Querying 'terrestrial-protected-areas-representation-by-biogeoclimatic-unit' record
#> ● Using collect() on this object will return 15267 features and 25 fields
#> ● At most six rows of the record are printed here
#> ─────────────────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 25 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 1084136 ymin: 465827.5 xmax: 1565462 ymax: 749984.1
#> epsg (SRID):    3005
#> proj4string:    +proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 6 x 26
#>   id    BG_CODE BCAREAPR PRPA_BC NUMPA_BC PCTBCPR PCTPPABC FEATURE_CLASS_S…
#>   <chr> <chr>      <dbl>   <dbl>    <int>   <dbl>    <dbl>            <int>
#> 1 WHSE… IDF dm…  155206.   5097.        9    3.28   0.0358              435
#> 2 WHSE… IDF dm…  155206.   5097.        9    3.28   0.0358              435
#> 3 WHSE… IDF dm…  155206.   5097.        9    3.28   0.0358              435
#> 4 WHSE… IDF dm…  155206.   5097.        9    3.28   0.0358              435
#> 5 WHSE… IDF dw    88067.  30714.        4   34.9    0.216               435
#> 6 WHSE… IDF dw    88067.  30714.        4   34.9    0.216               435
#> # … with 18 more variables: ZONE <chr>, SUBZONE <chr>, VARIANT <chr>,
#> #   PHASE <chr>, NATURAL_DISTURBANCE <chr>, MAP_LABEL <chr>, ZONE_NAME <chr>,
#> #   SUBZONE_NAME <chr>, VARIANT_NAME <chr>, PHASE_NAME <chr>,
#> #   NATURAL_DISTURBANCE_NAME <chr>, FEATURE_AREA <int>, FEATURE_LENGTH <int>,
#> #   FEATURE_AREA_SQM <dbl>, FEATURE_LENGTH_M <dbl>, OBJECTID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POLYGON [m]>

## Using a BCGW name
bcdc_query_geodata("WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW")

#> Querying 'WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW' record
#> ● Using collect() on this object will return 455 features and 41 fields
#> ● At most six rows of the record are printed here
#> ─────────────────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 41 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 833323.9 ymin: 381604.1 xmax: 1198292 ymax: 1054950
#> epsg (SRID):    3005
#> proj4string:    +proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 6 x 42
#>   id    CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>   <chr> <chr>            <chr>            <chr>            <chr>           
#> 1 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 2 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 3 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 4 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 5 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> 6 WHSE… "Ministry of Fo… airTransportati… Air Transportat… BC Airports     
#> # … with 37 more variables: SOURCE_DATA_ID <chr>, SUPPLIED_SOURCE_ID_IND <chr>,
#> #   AIRPORT_NAME <chr>, DESCRIPTION <chr>, PHYSICAL_ADDRESS <chr>,
#> #   ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>, POSTAL_CODE <chr>,
#> #   LOCALITY <chr>, CONTACT_PHONE <chr>, CONTACT_EMAIL <chr>,
#> #   CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>, LATITUDE <dbl>,
#> #   LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <dbl>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

# }