Query data from the B.C. Web Feature Service

Queries features from the B.C. Web Feature Service. See bcdc_tidy_resources() - if a resource has a value of "wms" in the format column it is available as a Web Feature Service, and you can query and download it using bcdc_query_geodata(). The response will be paginated if the number of features is above the number set by the bcdata.single_download_limit option. Please see bcdc_options() for defaults and more information.

Usage

bcdc_query_geodata(record, crs = 3005)

Arguments

record

either a bcdc_record object (from the result of bcdc_get_record()), a character string denoting the name or ID of a resource (or the URL) or a BC Geographic Warehouse (BCGW) name.

It is advised to use the permanent ID for a record or the BCGW name rather than the human-readable name to guard against future name changes of the record. If you use the human-readable name a warning will be issued once per session. You can silence these warnings altogether by setting an option: options("silence_named_get_data_warning" = TRUE) - which you can set in your .Rprofile file so the option persists across sessions.

crs

the epsg code for the coordinate reference system. Defaults to 3005 (B.C. Albers). See https://epsg.io.

Value

A bcdc_promise object. This object includes all of the information required to retrieve the requested data. In order to get the actual data as an sf object, you need to run collect() on the bcdc_promise.

Details

Note that this function doesn't actually return the data, but rather an object of class bcdc_promise, which includes all of the information required to retrieve the requested data. In order to get the actual data as an sf object, you need to run collect() on the bcdc_promise. This allows further refining the call to bcdc_query_geodata() with filter() and/or select() statements before pulling down the actual data as an sf object with collect(). See examples.

Examples

# \donttest{
# Returns a bcdc_promise, which can be further refined using filter/select:
try(
  bcdc_query_geodata("bc-airports", crs = 3857)
)
#> Querying 'bc-airports' record
#> • Using collect() on this object will return
#> • 455 features and 41 fields
#> • At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 41 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -14313270 ymin: 6176688 xmax: -13728510 ymax: 7259410
#> Projected CRS: WGS 84 / Pseudo-Mercator
#> # A tibble: 6 × 42
#>   id      CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO…
#>   <chr>   <chr>            <chr>            <chr>           
#> 1 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 2 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 3 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 4 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 5 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 6 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> # … with 38 more variables:
#> #   OCCUPANT_TYPE_DESCRIPTION <chr>, SOURCE_DATA_ID <chr>,
#> #   SUPPLIED_SOURCE_ID_IND <chr>, AIRPORT_NAME <chr>,
#> #   DESCRIPTION <chr>, PHYSICAL_ADDRESS <chr>,
#> #   ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>,
#> #   POSTAL_CODE <chr>, LOCALITY <chr>, CONTACT_PHONE <chr>,
#> #   CONTACT_EMAIL <chr>, CONTACT_FAX <chr>, …

# To obtain the actual data as an sf object, collect() must be called:
try(
  bcdc_query_geodata("bc-airports", crs = 3857) %>%
    filter(PHYSICAL_ADDRESS == 'Victoria, BC') %>%
    collect()
)
#> Note: method with signature ‘DBIConnection#character’ chosen for function ‘dbQuoteIdentifier’,
#>  target signature ‘wfsConnection#ident’.
#>  "wfsConnection#ANY" would also be valid
#> Note: method with signature ‘DBIConnection#character’ chosen for function ‘dbQuoteIdentifier’,
#>  target signature ‘wfsConnection#character’.
#>  "wfsConnection#ANY" would also be valid
#> Note: method with signature ‘DBIConnection#character’ chosen for function ‘dbQuoteString’,
#>  target signature ‘wfsConnection#character’.
#>  "wfsConnection#ANY" would also be valid
#> Simple feature collection with 3 features and 41 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -13735460 ymin: 6176688 xmax: -13733670 ymax: 6177610
#> Projected CRS: WGS 84 / Pseudo-Mercator
#> # A tibble: 3 × 42
#>   id      CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO…
#> * <chr>   <chr>            <chr>            <chr>           
#> 1 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 2 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 3 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> # … with 38 more variables:
#> #   OCCUPANT_TYPE_DESCRIPTION <chr>, SOURCE_DATA_ID <chr>,
#> #   SUPPLIED_SOURCE_ID_IND <chr>, AIRPORT_NAME <chr>,
#> #   DESCRIPTION <chr>, PHYSICAL_ADDRESS <chr>,
#> #   ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>,
#> #   POSTAL_CODE <chr>, LOCALITY <chr>, CONTACT_PHONE <chr>,
#> #   CONTACT_EMAIL <chr>, CONTACT_FAX <chr>, …

try(
  bcdc_query_geodata("groundwater-wells") %>%
    filter(OBSERVATION_WELL_NUMBER == "108") %>%
    select(WELL_TAG_NUMBER, INTENDED_WATER_USE) %>%
    collect()
)
#> Simple feature collection with 1 feature and 5 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 1273428 ymin: 1197388 xmax: 1273428 ymax: 1197388
#> Projected CRS: NAD83 / BC Albers
#> # A tibble: 1 × 6
#>   id            GW_WW_SYSID WELL_TAG_NUMBER INTENDED_WATER_…
#> * <chr>               <int>           <int> <chr>           
#> 1 WHSE_WATER_M…       22688           22722 Not Applicable  
#> # … with 2 more variables: OBJECTID <int>,
#> #   geometry <POINT [m]>

## A moderately large layer
try(
  bcdc_query_geodata("bc-environmental-monitoring-locations")
)
#> Querying
#>  'bc-environmental-monitoring-locations'
#>  record
#> • Using collect() on this object will return
#> • 38778 features and 35 fields
#> • Accessing this record requires pagination and will
#> • make 39 separate requests to the WFS. See
#> • ?bcdc_options
#> • At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 35 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 1208927 ymin: 456538.9 xmax: 1297743 ymax: 481711.2
#> Projected CRS: NAD83 / BC Albers
#> # A tibble: 6 × 36
#>   id      MONITORING_LOCA… CLIENT_LOCATION… CONTAMINATED_SI…
#>   <chr>   <chr>                       <int> <chr>           
#> 1 WHSE_E… 0301437                        20 NA              
#> 2 WHSE_E… 0301438                        20 NA              
#> 3 WHSE_E… 0301439                        20 NA              
#> 4 WHSE_E… 0301454                        20 NA              
#> 5 WHSE_E… 0301461                        20 NA              
#> 6 WHSE_E… 0301462                        20 NA              
#> # … with 32 more variables: DISCHARGE_MEDIA_CD <chr>,
#> #   ECO_REGION_CD <chr>, GEOREFERENCE_SOURCE_CD <chr>,
#> #   LOCATION_PURPOSE_CD <chr>, LOCATION_TYPE_CD <chr>,
#> #   OFFICE_CD <chr>, SOURCE_CLASS_CD <chr>,
#> #   WATERSHED_CD <chr>, CALC_STATISTICS_IND <chr>,
#> #   LOCATION_AREA_IND <chr>, MOBILE_LOCATION_IND <chr>,
#> #   CLOSED_DATE <chr>, DESCRIPTION <chr>, …

try(
  bcdc_query_geodata("bc-environmental-monitoring-locations") %>%
    filter(PERMIT_RELATIONSHIP == "DISCHARGE")
)
#> Querying
#>  'bc-environmental-monitoring-locations'
#>  record
#> • Using collect() on this object will return
#> • 12269 features and 35 fields
#> • Accessing this record requires pagination and will
#> • make 13 separate requests to the WFS. See
#> • ?bcdc_options
#> • At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 35 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 887419.7 ymin: 431835.5 xmax: 1182601 ymax: 621133.7
#> Projected CRS: NAD83 / BC Albers
#> # A tibble: 6 × 36
#>   id      MONITORING_LOCA… CLIENT_LOCATION… CONTAMINATED_SI…
#>   <chr>   <chr>                       <int> <chr>           
#> 1 WHSE_E… 0100050                        11 NA              
#> 2 WHSE_E… 0140200                        11 NA              
#> 3 WHSE_E… 0152664                        11 NA              
#> 4 WHSE_E… 0160019                        11 NA              
#> 5 WHSE_E… 0160022                        11 NA              
#> 6 WHSE_E… 0160025                        11 NA              
#> # … with 32 more variables: DISCHARGE_MEDIA_CD <chr>,
#> #   ECO_REGION_CD <chr>, GEOREFERENCE_SOURCE_CD <chr>,
#> #   LOCATION_PURPOSE_CD <chr>, LOCATION_TYPE_CD <chr>,
#> #   OFFICE_CD <chr>, SOURCE_CLASS_CD <chr>,
#> #   WATERSHED_CD <chr>, CALC_STATISTICS_IND <chr>,
#> #   LOCATION_AREA_IND <chr>, MOBILE_LOCATION_IND <chr>,
#> #   CLOSED_DATE <chr>, DESCRIPTION <chr>, …

## A very large layer
try(
  bcdc_query_geodata("terrestrial-protected-areas-representation-by-biogeoclimatic-unit")
)
#> Querying
#>  'terrestrial-protected-areas-representation-by-biogeoclimatic-unit'
#>  record
#> • Using collect() on this object will return
#> • 15666 features and 25 fields
#> • Accessing this record requires pagination and will
#> • make 16 separate requests to the WFS. See
#> • ?bcdc_options
#> • At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 25 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 381212.5 ymin: 1183638 xmax: 965704.6 ymax: 1651988
#> Projected CRS: NAD83 / BC Albers
#> # A tibble: 6 × 26
#>   id      BG_CODE BCAREAPR PRPA_BC NUMPA_BC PCTBCPR PCTPPABC
#>   <chr>   <chr>      <dbl>   <dbl>    <int>   <dbl>    <dbl>
#> 1 WHSE_L… BAFAun  6069670.  1.70e6       75    28.0     12.8
#> 2 WHSE_L… BAFAun  6069670.  1.70e6       75    28.0     12.8
#> 3 WHSE_L… BAFAun  6069670.  1.70e6       75    28.0     12.8
#> 4 WHSE_L… BAFAun  6069670.  1.70e6       75    28.0     12.8
#> 5 WHSE_L… BAFAun  6069670.  1.70e6       75    28.0     12.8
#> 6 WHSE_L… BAFAun  6069670.  1.70e6       75    28.0     12.8
#> # … with 19 more variables: FEATURE_CLASS_SKEY <int>,
#> #   ZONE <chr>, SUBZONE <chr>, VARIANT <chr>, PHASE <chr>,
#> #   NATURAL_DISTURBANCE <chr>, MAP_LABEL <chr>,
#> #   ZONE_NAME <chr>, SUBZONE_NAME <chr>,
#> #   VARIANT_NAME <chr>, PHASE_NAME <chr>,
#> #   NATURAL_DISTURBANCE_NAME <chr>, FEATURE_AREA <int>,
#> #   FEATURE_LENGTH <int>, FEATURE_AREA_SQM <dbl>, …

## Using a BCGW name
try(
  bcdc_query_geodata("WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW")
)
#> Querying
#>  'WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW'
#>  record
#> • Using collect() on this object will return
#> • 455 features and 41 fields
#> • At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 41 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 833323.9 ymin: 381604.1 xmax: 1198292 ymax: 1054950
#> Projected CRS: NAD83 / BC Albers
#> # A tibble: 6 × 42
#>   id      CUSTODIAN_ORG_D… BUSINESS_CATEGO… BUSINESS_CATEGO…
#>   <chr>   <chr>            <chr>            <chr>           
#> 1 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 2 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 3 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 4 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 5 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> 6 WHSE_I… "Ministry of Fo… airTransportati… Air Transportat…
#> # … with 38 more variables:
#> #   OCCUPANT_TYPE_DESCRIPTION <chr>, SOURCE_DATA_ID <chr>,
#> #   SUPPLIED_SOURCE_ID_IND <chr>, AIRPORT_NAME <chr>,
#> #   DESCRIPTION <chr>, PHYSICAL_ADDRESS <chr>,
#> #   ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>,
#> #   POSTAL_CODE <chr>, LOCALITY <chr>, CONTACT_PHONE <chr>,
#> #   CONTACT_EMAIL <chr>, CONTACT_FAX <chr>, …
# }