Search the B.C. Data Catalogue

Search the B.C. Data Catalogue

bcdc_search(
  ...,
  license_id = NULL,
  download_audience = "Public",
  res_format = NULL,
  sector = NULL,
  organization = NULL,
  n = 100
)

Arguments

...
search terms

license_id
the type of license (see bcdc_search_facets("license_id")).

download_audience
download audience (see bcdc_search_facets("download_audience")). Default "Public"

res_format
format of resource (see bcdc_search_facets("res_format"))

sector
sector of government from which the data comes (see bcdc_search_facets("sector"))

organization
government organization that manages the data (see bcdc_search_facets("organization"))

n
number of results to return. Default 100

Value

A list containing the records that match the search

Examples

# \donttest{
try(
  bcdc_search("forest")
)
#> Found 275 matches. Returning the first 100.
#> To see them all, rerun the search and set the 'n' argument to 275.
#> List of B.C. Data Catalogue Records
#> Number of records: 100
#> Showing the top 50 results. You can assign the output of
#>  bcdc_search, to an object and subset with `[` to see other results in
#>  the set.
#> 
#> Titles:
#> 1: Forest Development Units (other, wms, kml)
#>  ID: 25bbc9a6-f0eb-47dd-849c-3161a6512468
#>  Name: forest-development-units
#> 2: Harvested Areas of BC (Consolidated Cutblocks) (other, fgdb, wms,
#>  kml, pdf)
#>  ID: b1b647a6-f271-42e0-9cd0-89ec24bce9f7
#>  Name: harvested-areas-of-bc-consolidated-cutblocks-
#> 3: RESULTS - Forest Cover Silviculture (other, wms, kml)
#>  ID: 258bb088-4113-47b1-b568-ce20bd64e3e3
#>  Name: results-forest-cover-silviculture
#> 4: RESULTS - Forest Cover Inventory (other, wms)
#>  ID: 56ac43a7-724a-4f01-b193-d5f9a16ef0a8
#>  Name: results-forest-cover-inventory
#> 5: Forest Tenure Managed Licence (other, wms, kml)
#>  ID: c3e96239-cdc9-4328-ac19-58fba1623ef8
#>  Name: forest-tenure-managed-licence
#> 6: Pest Infestation Polygons (other, zip)
#>  ID: 450b67bb-02d5-4526-8bc0-ac7924125a1e
#>  Name: pest-infestation-polygons
#> 7: Pest Infestation Points (other)
#>  ID: 4a83ec61-31a4-4134-9204-af47ff26ed78
#>  Name: pest-infestation-points
#> 8: Forest Tenure Harvesting Authority Polygons (other, wms, kml)
#>  ID: cff7b8f7-6897-444f-8c53-4bb93c7e9f8b
#>  Name: forest-tenure-harvesting-authority-polygons
#> 9: Generalized Forest Cover Ownership (wms, kml, other)
#>  ID: 5fc4e8ce-dd1d-44fd-af17-e0789cf65e4e
#>  Name: generalized-forest-cover-ownership
#> 10: Forest Tenure Cutblock Polygons (FTA 4.0) (other, wms, kml)
#>  ID: dfb8b498-fa4b-4286-b3ec-58db88aca1cf
#>  Name: forest-tenure-cutblock-polygons-fta-4-0
#> 11: FDU - Forest Licence Reference (other)
#>  ID: 7e986391-c740-4d1b-9294-7576af2d2998
#>  Name: fdu-forest-licence-reference
#> 12: Forest Tenure Road Section Lines (other, wms, kml)
#>  ID: 243c94a1-f275-41dc-bc37-91d8a2b26e10
#>  Name: forest-tenure-road-section-lines
#> 13: RESULTS - Forest Cover Reserve (other, wms, kml)
#>  ID: 7028ac47-45dd-41d7-a371-be5a04177afe
#>  Name: results-forest-cover-reserve
#> 14: Forest Tenure Real Property Project (other, wms, kml)
#>  ID: 3d7f09c3-ff25-429e-a4d9-d88f9919d904
#>  Name: forest-tenure-real-property-project
#> 15: Pest Infestation Polygons - Historic (other)
#>  ID: b6592d4d-1802-469e-840f-347baae24726
#>  Name: pest-infestation-polygons-historic
#> 16: Pest Infestation Points - Historic (other)
#>  ID: 3a7b4e9d-2761-4e8f-a8ec-a79265288e7b
#>  Name: pest-infestation-points-historic
#> 17: Forest Tenure Map Notation Point (other, wms)
#>  ID: 69f4f070-d70a-43c8-b224-354758f4f492
#>  Name: forest-tenure-map-notation-point
#> 18: Pest Infestation Polygons - Current (other, wms, kml)
#>  ID: 967c4d29-817a-418d-b933-359c311a30a8
#>  Name: pest-infestation-polygons-current
#> 19: Pest Infestation Points - Current (other, wms, kml)
#>  ID: 1f885327-eb89-4905-84be-8d9710641648
#>  Name: pest-infestation-points-current
#> 20: Forest Tenure Map Notation Polygon (other, wms, kml)
#>  ID: 78838b54-f241-4269-aeb4-126d14bb5f37
#>  Name: forest-tenure-map-notation-polygon
#> 21: FADM - Provincial Forest Deletion (other, wms, kml)
#>  ID: fc480dec-1e67-4944-9343-6c72ae124440
#>  Name: fadm-provincial-forest-deletion
#> 22: FADM - Provincial Forest Addition (other, wms, kml)
#>  ID: 07dba191-3ebc-4c32-af7e-16769c237011
#>  Name: fadm-provincial-forest-addition
#> 23: Forest Tenure Map Notation Line (other, wms, kml)
#>  ID: 8ef18f74-e53f-497a-968d-86e9a20419a2
#>  Name: forest-tenure-map-notation-line
#> 24: Forest Tenure Free Use Permit (other, wms, kml)
#>  ID: 814e0ed9-95c3-4750-8152-1307795986f1
#>  Name: forest-tenure-free-use-permit
#> 25: Forest Tenure Special Access Road Polygon (other, wms, kml)
#>  ID: 6d171eb8-d7a8-4ad5-8a9e-027e43f7a54b
#>  Name: forest-tenure-special-access-road-polygon
#> 26: Forest Tenure Special Access Road Line (other, wms, kml)
#>  ID: 4ee37a5d-6a14-4f98-8ad9-566b4ea28e2b
#>  Name: forest-tenure-special-access-road-line
#> 27: Forest Tenure Road Segment Lines (other, wms, kml)
#>  ID: 9e5bfa62-2339-445e-bf67-81657180c682
#>  Name: forest-tenure-road-segment-lines
#> 28: Forest Tenure Timber Licence Elimination (other, wms, kml)
#>  ID: e15e0543-830f-43b8-b824-ed9a8ed7eeca
#>  Name: forest-tenure-timber-licence-elimination
#> 29: Forest Tenure Timber Licence (other, wms, kml)
#>  ID: f2356651-4456-4d93-ba17-7b8c4a3ee53d
#>  Name: forest-tenure-timber-licence
#> 30: Recreation Sites, Reserves, and Interpretive Forests Details and
#>  Closures (wms, kml, other)
#>  ID: b0f23ae8-e0aa-449c-a806-653ba37f100a
#>  Name:
#>   recreation-sites-reserves-and-interpretive-forests-details-and-closures
#> 31: Forest Tenure Communication Sites (other, wms, kml)
#>  ID: 1361d483-26ae-4ca9-adba-e19c3efac147
#>  Name: forest-tenure-communication-sites
#> 32: FADM - Provincial Forest Exclusion (other, wms, kml)
#>  ID: f3239e0c-f347-49f9-a53d-39290e597159
#>  Name: fadm-provincial-forest-exclusion
#> 33: FADM - Provincial Forest (other, wms, kml)
#>  ID: 4a023cb2-09fd-44c2-b3ce-439ecb5f39f3
#>  Name: fadm-provincial-forest
#> 34: RESULTS - Forest Cover Species (Attribute Only) (other)
#>  ID: 2db77f8b-765b-488f-9752-f1349cf6ebfc
#>  Name: results-forest-cover-species-attribute-only
#> 35: Productive Forest Landbase for the Cariboo Region (other, wms, kml)
#>  ID: 0cfa7f53-272d-47d4-83df-95b4edd75460
#>  Name: productive-forest-landbase-for-the-cariboo-region
#> 36: Forest Tenure Special Use Permit Polygon (other, wms, kml)
#>  ID: d29b37fc-cbfe-4ebf-ac6e-2bf8fa926a81
#>  Name: forest-tenure-special-use-permit-polygon
#> 37: VRI - 2020 - Forest Vegetation Composite Polygons (fgdb)
#>  ID: 6ba30649-14cd-44ad-a11f-794feed39f40
#>  Name: vri-2020-forest-vegetation-composite-polygons
#> 38: Forest Licensee Operating Areas (other, wms, kml)
#>  ID: 5a18c401-7c99-48fe-8a1a-b4597072d8b6
#>  Name: forest-licensee-operating-areas
#> 39: Forest Tenure Road Section Amendment (other)
#>  ID: bad79704-3fc9-43e7-b5d4-660657199af8
#>  Name: forest-tenure-road-section-amendment
#> 40: VRI - 2020 - Variable Density Yield Projection 7 (VDYP7) Input
#>  Polygon (other)
#>  ID: 57513aaa-c0a6-41a9-b2a8-b980b1604ee6
#>  Name: vri-2020-variable-density-yield-projection-7-vdyp7-input-polygon
#> 41: Recreation Polygons (other, wms, kml)
#>  ID: 263338a7-93ee-49c1-83e8-13f0bde70833
#>  Name: recreation-polygons
#> 42: Recreation Sites Subset - Information Purposes Only (other, wms,
#>  kml)
#>  ID: bc37b35d-8d00-45ab-9ab2-57cae75637ad
#>  Name: recreation-sites-subset-information-purposes-only
#> 43: Forest Inventory Zones (other, wms, kml)
#>  ID: 67e95c68-c1ef-4363-b351-0dfead151122
#>  Name: forest-inventory-zones
#> 44: Forest Tenure Timber Licence Remaining (other)
#>  ID: edf864e8-bae6-4735-a85b-3e6f276d4a14
#>  Name: forest-tenure-timber-licence-remaining
#> 45: Mixed Forest 1:250,000 GeoBase Land Cover (other)
#>  ID: 902f5af2-9fce-4579-b48d-8bd7f85d36dd
#>  Name: mixed-forest-1-250-000-geobase-land-cover
#> 46: Coniferous Forest 1:250,000 GeoBase Land Cover (other)
#>  ID: 177b5ef0-8339-4fd7-a40a-93edb3797d07
#>  Name: coniferous-forest-1-250-000-geobase-land-cover
#> 47: Broadleaf Forest 1:250,000 GeoBase Land Cover (other)
#>  ID: 78bb0ee9-96ed-49fb-9c0c-7434b2f24d7e
#>  Name: broadleaf-forest-1-250-000-geobase-land-cover
#> 48: Seed Planning Units (Spatial View) (other, wms, kml)
#>  ID: 3d544ad5-8d1c-4cd9-a2db-05104d963145
#>  Name: seed-planning-units-spatial-view-
#> 49: Seedlot Spatial Area of Use (fgdb)
#>  ID: 46149c20-58f9-4447-8f4b-c52bb2333697
#>  Name: seedlot-spatial-area-of-use
#> 50: VRI - 2020 - Forest Vegetation Composite Rank 1 Layer (R1) (fgdb,
#>  other, wms, kml)
#>  ID: 2ebb35d8-c82f-4a17-9c96-612ac3532d55
#>  Name: vri-2020-forest-vegetation-composite-rank-1-layer-r1-
#> 
#> Access a single record by calling `bcdc_get_record(ID)` with the ID
#>  from the desired record.

try(
  bcdc_search("regional district", res_format = "fgdb")
)
#> List of B.C. Data Catalogue Records
#> Number of records: 95
#> Showing the top 50 results. You can assign the output of
#>  bcdc_search, to an object and subset with `[` to see other results in
#>  the set.
#> 
#> Titles:
#> 1: Regional District Boundaries - Road centreline aligned (other, wms,
#>  kml)
#>  ID: b66ed992-16c9-4527-b363-0ae4827d103b
#>  Name: regional-district-boundaries-road-centreline-aligned
#> 2: Regional district financial statistics (other)
#>  ID: 35fd9fe7-e937-4719-82e8-852cc699a113
#>  Name: regional-district-financial-statistics
#> 3: Regional Districts - Legally Defined Administrative Areas of BC
#>  (other, xlsx, wms, kml)
#>  ID: d1aff64e-dbfe-45a6-af97-582b7f6418b9
#>  Name: regional-districts-legally-defined-administrative-areas-of-bc
#> 4: Custom Census Reports (2016, 2011, 2006) (other, xlsx)
#>  ID: 894497d9-6901-4ed8-8655-79010e4b9b09
#>  Name: custom-census-reports-2016-2011-2006-
#> 5: CEEI Supporting Indicators Total 2009 Regional Districts (other)
#>  ID: 32966665-8e5c-4f9e-b10d-311355f2a1f0
#>  Name: ceei-supporting-indicators-total-2009-regional-districts
#> 6: CEEI Supporting Indicators Total 2006 Regional Districts (other)
#>  ID: d1be3579-82af-489b-90dd-3463c06dee58
#>  Name: ceei-supporting-indicators-total-2006-regional-districts
#> 7: CEEI Supporting Indicators Total 2001 Regional Districts (other)
#>  ID: 21486b7c-480a-4757-a7f8-8c780d156249
#>  Name: ceei-supporting-indicators-total-2001-regional-districts
#> 8: CEEI Supporting Indicators Total 1996 Regional Districts (other)
#>  ID: dd1c1225-3a4c-4181-85d2-2c07f05a91ca
#>  Name: ceei-supporting-indicators-total-1996-regional-districts
#> 9: Recreation Sites and Trails BC District Boundaries (fgdb, other,
#>  wms, kml)
#>  ID: 0ad49841-a25b-4be4-b98c-9e817e94cc48
#>  Name: recreation-sites-and-trails-bc-district-boundaries
#> 10: Forest Licensee Operating Areas (other, wms, kml)
#>  ID: 5a18c401-7c99-48fe-8a1a-b4597072d8b6
#>  Name: forest-licensee-operating-areas
#> 11: Legally Defined Administrative Areas of BC (other, xlsx)
#>  ID: 12f0fd64-94d5-422a-832e-6627a2a01702
#>  Name: legally-defined-administrative-areas-of-bc
#> 12: Population counts by age by Census Division (Regional District)-
#>  2016 (other, wms, kml)
#>  ID: 348c21df-78b9-4ead-9ac3-7fd8703a7c36
#>  Name:
#>   population-counts-by-age-by-census-division-regional-district-2016
#> 13: Recreation Sites and Trail BC Region Boundaries (fgdb, other, wms,
#>  kml)
#>  ID: 919f236a-e40f-4f07-a459-f976f02b6e11
#>  Name: recreation-sites-and-trail-bc-region-boundaries
#> 14: CEEI Primary Indicators Per Capita 2007 Regional Districts (other,
#>  wms, kml)
#>  ID: fb979ed6-9d05-4de8-b1b6-f8acb212bfcb
#>  Name: ceei-primary-indicators-per-capita-2007-regional-districts
#> 15: CEEI Primary Indicators Total 2007 Regional Districts (other, wms,
#>  kml)
#>  ID: 05a664ce-c255-43dd-98a0-ea9735d08f8f
#>  Name: ceei-primary-indicators-total-2007-regional-districts
#> 16: Enhanced Riparian Zones LRMP Lillooet Forest District - Polys
#>  (other)
#>  ID: 06b9189a-8c68-4a22-b007-ae6b9156abc3
#>  Name: enhanced-riparian-zones-lrmp-lillooet-forest-district-polys
#> 17: North Thompson Caribou Habitat Clearwater Forest District (other)
#>  ID: 6271d5dc-4870-41e4-903d-4d378c297353
#>  Name: north-thompson-caribou-habitat-clearwater-forest-district
#> 18: Biodiversity Corridors Kamloops Forest District (other)
#>  ID: f15d11d3-2a3e-41a5-a576-dab136cced9b
#>  Name: biodiversity-corridors-kamloops-forest-district
#> 19: Legally Defined Administrative Areas of BC Boundary Locations
#>  (other, xlsx, wms, kml)
#>  ID: e50b6560-0549-41a6-8035-6b828066692e
#>  Name: legally-defined-administrative-areas-of-bc-boundary-locations
#> 20: Natural Resource (NR) Districts (other, wms)
#>  ID: 0bc73892-e41f-41d0-8d8e-828c16139337
#>  Name: natural-resource-nr-district
#> 21: Ministry of Transportation (MOT) District Boundary (other, wms,
#>  kml)
#>  ID: ee6102a7-dd95-4746-aeca-874e717f238a
#>  Name: ministry-of-transportation-mot-district-boundary
#> 22: Rose Swanson Sensitive Area Vernon Forest District (other)
#>  ID: 30a88772-f815-4795-8f4d-9c09d52935a9
#>  Name: rose-swanson-sensitive-area-vernon-forest-district
#> 23: Caribou Movement Corridors Clearwater Forest District - Lines
#>  (other)
#>  ID: 7c7768f2-7a3d-4a8f-958e-e7a8dd2b18f2
#>  Name: caribou-movement-corridors-clearwater-forest-district-lines
#> 24: Recreation Trails for the Cariboo Region (other)
#>  ID: e919372d-5fa1-4214-9f57-1d606a06f1f4
#>  Name: recreation-trails-for-the-cariboo-region
#> 25: BC College Region Boundaries (other, wms, kml)
#>  ID: 6b8da230-9bc2-4bb4-989e-7d5dad84a4ac
#>  Name: bc-college-region-boundaries
#> 26: Walk In Lakes and Surrounding Buffers Clearwater FD (other)
#>  ID: 68110f99-c5ce-47da-8cf8-472ae100d2e4
#>  Name: walk-in-lakes-and-surrounding-buffers-clearwater-fd
#> 27: TANTALIS - Land Districts (other, wms, kml)
#>  ID: fe1023fc-5c8d-49ec-8f62-86fd16649ba2
#>  Name: tantalis-land-districts
#> 28: Mountain Goat Inventory Areas for the Mid-Coast Forest District
#>  (other)
#>  ID: 172a4ff9-9257-4a30-9c87-392ecc0a1b53
#>  Name: mountain-goat-inventory-areas-for-the-mid-coast-forest-district
#> 29: Operability - Revelstoke TSA (other)
#>  ID: 785c9d5a-f435-493f-9ec5-bc2ac618ad0a
#>  Name: operability-revelstoke-tsa
#> 30: Potential Moose Habitat Lillooet Forest District (other)
#>  ID: de3be33c-18b1-40e7-a61c-70784ad20a15
#>  Name: potential-moose-habitat-lillooet-forest-district
#> 31: Operability - Golden TSA (other)
#>  ID: e6d68f79-4df4-4089-96c6-a7d87a5d16a3
#>  Name: operability-golden-tsa
#> 32: Operability - Cranbrook TSA (other)
#>  ID: 07c1e904-7541-4b56-87e4-0435bbc82a43
#>  Name: operability-cranbrook-tsa
#> 33: Terrestrial Ecosystem Mapping - Steam Boat Mtn (other)
#>  ID: 0cca68e7-50b1-4f4c-a28b-0e5ce9d73278
#>  Name: terrestrial-ecosystem-mapping-steam-boat-mtn
#> 34: Terrestrial Ecosystem Mapping - Sheep Mtn (other)
#>  ID: 09a6f1c7-2f48-40dd-9a44-e4df72d35c4c
#>  Name: terrestrial-ecosystem-mapping-sheep-mtn
#> 35: Terrestrial Ecosystem Mapping - Fort Sheppard (other)
#>  ID: 5b75978f-02b9-457b-9dd4-71f0927bf7dc
#>  Name: terrestrial-ecosystem-mapping-fort-sheppard
#> 36: Terrestrial Ecosystem Mapping - East Columbia Lake (other)
#>  ID: 1e1a6668-3a97-43ba-9875-f2ea6d4deb89
#>  Name: terrestrial-ecosystem-mapping-east-columbia-lake
#> 37: Terrestrial Ecosystem Mapping - Cummins River (other)
#>  ID: a7ed08c3-3479-41e8-8a24-583297cbc437
#>  Name: terrestrial-ecosystem-mapping-cummins-river
#> 38: Terrestrial Ecosystem Mapping - Bull Mtn (other)
#>  ID: 477cfc3c-f2af-4e50-a2ec-18015606e5fb
#>  Name: terrestrial-ecosystem-mapping-bull-mtn
#> 39: Terrestrial Ecosystem Mapping - Broadwater (other)
#>  ID: ea19e0a1-a8f7-4299-829e-1faf19540716
#>  Name: terrestrial-ecosystem-mapping-broadwater
#> 40: Terrestrial Ecosystem Mapping - Brewer Cr (other)
#>  ID: 105168de-7be4-41f4-9151-276a4ed32b27
#>  Name: terrestrial-ecosystem-mapping-brewer-cr
#> 41: Grizzly Class 1 and 2 Habitat for the Mid-Coast Forest District
#>  (other)
#>  ID: a6d7c440-d765-4f8f-833c-6530b29d32f5
#>  Name: grizzly-class-1-and-2-habitat-for-the-mid-coast-forest-district
#> 42: Current Provincial Electoral Districts of British Columbia (other,
#>  wms, kml)
#>  ID: 050ec630-47dd-476b-b773-28b2f8d8bc98
#>  Name: current-provincial-electoral-districts-of-british-columbia
#> 43: Local and Regional Greenspaces (wms, kml, xlsx, other)
#>  ID: 6a2fea1b-0cc4-4fc2-8017-eaf755d516da
#>  Name: local-and-regional-greenspaces
#> 44: Terrestrial Ecosystem Mapping - Wilson/Stagleap Creeks (other)
#>  ID: 9ec86937-b662-449e-b956-00c1e126843e
#>  Name: terrestrial-ecosystem-mapping-wilson-stagleap-creeks
#> 45: Terrestrial Ecosystem Mapping - TFL 14 / Spillimachine (other)
#>  ID: 9fcc69e3-8f80-44f3-99ad-0e04cb9fa52d
#>  Name: terrestrial-ecosystem-mapping-tfl-14-spillimachine
#> 46: Terrestrial Ecosystem Mapping - Stoddart Creek (other)
#>  ID: de18eee8-8e66-4dcb-8e07-ce8e1a2b02de
#>  Name: terrestrial-ecosystem-mapping-stoddart-creek
#> 47: Terrestrial Ecosystem Mapping - MT Broadwood Heritage Cons Area
#>  (other)
#>  ID: c22d3bd9-a307-4666-be26-ebc82890ce62
#>  Name: terrestrial-ecosystem-mapping-mt-broadwood-heritage-cons-area
#> 48: Terrestrial Ecosystem Mapping - Fort Sheppard - Sample Plots
#>  (other)
#>  ID: ac50baf2-17cf-427a-9fb2-b5b42ec30bb3
#>  Name: terrestrial-ecosystem-mapping-fort-sheppard-sample-plots
#> 49: Operability - Boundary TSA (other)
#>  ID: 7c1fb96f-6e1a-40cb-8f27-0c3892a91356
#>  Name: operability-boundary-tsa
#> 50: Provincial Electoral Districts - Electoral Boundaries
#>  Redistribution, 2015 (other, kmz, shp, wms, kml)
#>  ID: 9530a41d-6484-41e5-b694-acb76e212a58
#>  Name:
#>   provincial-electoral-districts-electoral-boundaries-redistribution-2015
#> 
#> Access a single record by calling `bcdc_get_record(ID)` with the ID
#>  from the desired record.
# }