Search the B.C. Data Catalogue

Search the B.C. Data Catalogue

bcdc_search(
  ...,
  license_id = NULL,
  download_audience = "Public",
  res_format = NULL,
  sector = NULL,
  organization = NULL,
  n = 100
)

Arguments

...

search terms

license_id

the type of license (see bcdc_search_facets("license_id")).

download_audience

download audience (see bcdc_search_facets("download_audience")). Default "Public"

res_format

format of resource (see bcdc_search_facets("res_format"))

sector

sector of government from which the data comes (see bcdc_search_facets("sector"))

organization

government organization that manages the data (see bcdc_search_facets("organization"))

n

number of results to return. Default 100

Value

A list containing the records that match the search

Examples

# \donttest{
try(
  bcdc_search("forest")
)
#> Found 288 matches. Returning the first 100.
#> To see them all, rerun the search and set the 'n' argument to 288.
#> List of B.C. Data Catalogue Records
#> Number of records: 100
#>  (Showing the top 10)
#> Titles:
#> 1: Forest Development Units (multiple, wms, kml)
#>  ID: 25bbc9a6-f0eb-47dd-849c-3161a6512468
#>  Name: forest-development-units
#> 2: FADM - Provincial Forest (multiple, wms, kml)
#>  ID: 4a023cb2-09fd-44c2-b3ce-439ecb5f39f3
#>  Name: fadm-provincial-forest
#> 3: Forest Inventory Zones (multiple, wms, kml)
#>  ID: 67e95c68-c1ef-4363-b351-0dfead151122
#>  Name: forest-inventory-zones
#> 4: Forest Tenure Managed Licence (multiple, wms, kml)
#>  ID: c3e96239-cdc9-4328-ac19-58fba1623ef8
#>  Name: forest-tenure-managed-licence
#> 5: FDU - Forest Licence Reference (multiple)
#>  ID: 7e986391-c740-4d1b-9294-7576af2d2998
#>  Name: fdu-forest-licence-reference
#> 6: FADM - Provincial Forest Deletion (multiple, wms, kml)
#>  ID: fc480dec-1e67-4944-9343-6c72ae124440
#>  Name: fadm-provincial-forest-deletion
#> 7: Forest Tenure Communication Sites (multiple, wms, kml)
#>  ID: 1361d483-26ae-4ca9-adba-e19c3efac147
#>  Name: forest-tenure-communication-sites
#> 8: Forest Tenure Timber Licence (multiple, wms, kml)
#>  ID: f2356651-4456-4d93-ba17-7b8c4a3ee53d
#>  Name: forest-tenure-timber-licence
#> 9: FADM - Provincial Forest Addition (multiple, wms, kml)
#>  ID: 07dba191-3ebc-4c32-af7e-16769c237011
#>  Name: fadm-provincial-forest-addition
#> 10: FADM - Provincial Forest Exclusion (multiple, wms, kml)
#>  ID: f3239e0c-f347-49f9-a53d-39290e597159
#>  Name: fadm-provincial-forest-exclusion
#> 
#> Access a single record by calling `bcdc_get_record(ID)` with the ID
#>  from the desired record.

try(
  bcdc_search("regional district", res_format = "fgdb")
)
#> List of B.C. Data Catalogue Records
#> Number of records: 19
#>  (Showing the top 10)
#> Titles:
#> 1: Recreation Trails for the Cariboo Region (other)
#>  ID: e919372d-5fa1-4214-9f57-1d606a06f1f4
#>  Name: recreation-trails-for-the-cariboo-region
#> 2: Recreation Sites and Trail BC Region Boundaries (fgdb, multiple,
#>  wms, kml)
#>  ID: 919f236a-e40f-4f07-a459-f976f02b6e11
#>  Name: recreation-sites-and-trail-bc-region-boundaries
#> 3: Grassland Benchmark for the Cariboo Region (other)
#>  ID: 2285c49d-a11d-4932-97d4-8c68695e03bb
#>  Name: grassland-benchmark-for-the-cariboo-region
#> 4: Recreation Sites and Trails BC District Boundaries (fgdb, multiple,
#>  wms, kml)
#>  ID: 0ad49841-a25b-4be4-b98c-9e817e94cc48
#>  Name: recreation-sites-and-trails-bc-district-boundaries
#> 5: Backcountry Recreation Areas for the Cariboo Region (other)
#>  ID: 95aaaa75-14a1-4afb-a5cd-e6b5bc8f2dd2
#>  Name: backcountry-recreation-areas-for-the-cariboo-region
#> 6: High Elevation Viewpoints for the Cariboo Region (other)
#>  ID: 76e34f98-55f7-461c-9843-30bfe7b72b95
#>  Name: high-elevation-viewpoints-for-the-cariboo-region
#> 7: Enhanced Riparian Zones LRMP Lillooet Forest District - Polys
#>  (other)
#>  ID: 06b9189a-8c68-4a22-b007-ae6b9156abc3
#>  Name: enhanced-riparian-zones-lrmp-lillooet-forest-district-polys
#> 8: Recreation Trails with Buffer for the Cariboo Region (other)
#>  ID: e403b4f8-1e78-4a8b-ac05-3d8c305d9f55
#>  Name: recreation-trails-with-buffer-for-the-cariboo-region
#> 9: Seral Stage Analysis 2004 Cariboo Region - Maps and Reports (other)
#>  ID: a4fd5e98-804c-4f95-888f-bedcc9964dbd
#>  Name: seral-stage-analysis-2004-cariboo-region-maps-and-reports
#> 10: Candidate Goal 2 Protected Areas for the Cariboo Region (other)
#>  ID: e67c7336-d9ac-4ca8-91ed-a512746a517e
#>  Name: candidate-goal-2-protected-areas-for-the-cariboo-region
#> 
#> Access a single record by calling `bcdc_get_record(ID)` with the ID
#>  from the desired record.
# }