Search the B.C. Data Catalogue

Search the B.C. Data Catalogue

Usage

bcdc_search(
  ...,
  license_id = NULL,
  download_audience = NULL,
  res_format = NULL,
  sector = NULL,
  organization = NULL,
  groups = NULL,
  n = 100
)

Arguments

...

search terms

license_id

the type of license (see bcdc_search_facets("license_id")).

download_audience

download audience (see bcdc_search_facets("download_audience")). Default NULL (all audiences).

res_format

format of resource (see bcdc_search_facets("res_format"))

sector

sector of government from which the data comes (see bcdc_search_facets("sector"))

organization

government organization that manages the data (see bcdc_search_facets("organization"))

groups

collections of datasets for a particular project or on a particular theme (see bcdc_search_facets("groups"))

n

number of results to return. Default 100

Value

A list containing the records that match the search

Examples

# \donttest{
try(
  bcdc_search("forest")
)
#> Found 567 matches. Returning the first 100.
#> To see them all, rerun the search and set the 'n' argument to 567.
#> List of B.C. Data Catalogue Records
#> Number of records: 100
#> Showing the top 50 results. You can assign the
#>  output of bcdc_search, to an object and subset with
#>  `[` to see other results in the set.
#> 
#> Titles:
#> 1: Non Replaceable Forest Licenses Kamloops Forest
#>  District (oracle_sde)
#>  ID: af8d871d-6612-4c94-a2ce-d05672ec9efa
#>  Name:
#>   non-replaceable-forest-licenses-kamloops-forest-district
#> 2: Forest Development Units (multiple, wms, kml,
#>  oracle_sde)
#>  ID: 25bbc9a6-f0eb-47dd-849c-3161a6512468
#>  Name: forest-development-units
#> 3: FADM - Provincial Forest (multiple, wms, kml)
#>  ID: 4a023cb2-09fd-44c2-b3ce-439ecb5f39f3
#>  Name: fadm-provincial-forest
#> 4: Forest Ecosystem Networks for the Chilcotin Forest
#>  District (oracle_sde)
#>  ID: e55ea13e-7503-4109-ae46-9c1a113eaed8
#>  Name:
#>   forest-ecosystem-networks-for-the-chilcotin-forest-district
#> 5: Forest Inventory Zones (multiple, wms, kml)
#>  ID: 67e95c68-c1ef-4363-b351-0dfead151122
#>  Name: forest-inventory-zones
#> 6: Forest MapView (other)
#>  ID: 67acfc63-1dea-43be-bee0-4bc8fa8901fc
#>  Name: forest-mapview
#> 7: Forest Tenure Managed Licence (multiple, wms, kml)
#>  ID: c3e96239-cdc9-4328-ac19-58fba1623ef8
#>  Name: forest-tenure-managed-licence
#> 8: Forest Tenure Road Lines (oracle_sde)
#>  ID: bed9be0f-3590-415b-b130-d5bfa2938604
#>  Name: forest-tenure-road-lines
#> 9: FDU - Forest Licence Reference (multiple)
#>  ID: 7e986391-c740-4d1b-9294-7576af2d2998
#>  Name: fdu-forest-licence-reference
#> 10: FADM - Provincial Forest Deletion (multiple, wms,
#>  kml)
#>  ID: fc480dec-1e67-4944-9343-6c72ae124440
#>  Name: fadm-provincial-forest-deletion
#> 11: FADM - Provincial Forest Addition (multiple, wms,
#>  kml)
#>  ID: 07dba191-3ebc-4c32-af7e-16769c237011
#>  Name: fadm-provincial-forest-addition
#> 12: Forest Tenure Communication Sites (multiple, wms,
#>  kml)
#>  ID: 1361d483-26ae-4ca9-adba-e19c3efac147
#>  Name: forest-tenure-communication-sites
#> 13: Forest Tenure Timber Licence (multiple, wms, kml)
#>  ID: f2356651-4456-4d93-ba17-7b8c4a3ee53d
#>  Name: forest-tenure-timber-licence
#> 14: FADM - Provincial Forest Exclusion (multiple,
#>  wms, kml)
#>  ID: f3239e0c-f347-49f9-a53d-39290e597159
#>  Name: fadm-provincial-forest-exclusion
#> 15: Forest Road Segment Tenure (multiple, wms, kml)
#>  ID: 07b0830e-a505-4c88-8d9a-8cf0a630f565
#>  Name: forest-road-segment-tenure
#> 16: Generalized Forest Harvesting Restrictions (fgdb,
#>  csv, gpkg)
#>  ID: 6c0bcb22-49cb-4315-bc9c-c4668495cc45
#>  Name: generalized-forest-harvesting-restrictions
#> 17: RESULTS - Forest Cover Silviculture (multiple,
#>  wms, kml)
#>  ID: 258bb088-4113-47b1-b568-ce20bd64e3e3
#>  Name: results-forest-cover-silviculture
#> 18: RESULTS - Forest Cover Inventory (multiple, wms)
#>  ID: 56ac43a7-724a-4f01-b193-d5f9a16ef0a8
#>  Name: results-forest-cover-inventory
#> 19: Generalized Forest Cover Ownership (wms, kml,
#>  multiple, pdf)
#>  ID: 5fc4e8ce-dd1d-44fd-af17-e0789cf65e4e
#>  Name: generalized-forest-cover-ownership
#> 20: RESULTS - Forest Cover Reserve (multiple, wms,
#>  kml)
#>  ID: 7028ac47-45dd-41d7-a371-be5a04177afe
#>  Name: results-forest-cover-reserve
#> 21: Forest Licensee Operating Areas (multiple, wms,
#>  kml)
#>  ID: 5a18c401-7c99-48fe-8a1a-b4597072d8b6
#>  Name: forest-licensee-operating-areas
#> 22: Forest Tenure Chart Block Polygons (oracle_sde)
#>  ID: 9d341916-d596-481c-b191-a0df41767c2f
#>  Name: forest-tenure-chart-block-polygons
#> 23: Forest Tenure Timber Use Polygons (oracle_sde)
#>  ID: 4b2afa31-a0b9-488f-8f15-1d6117e0093e
#>  Name: forest-tenure-timber-use-polygons
#> 24: Forest Tenure Cut Block Polygons (oracle_sde)
#>  ID: 214e60a2-4f1a-4a68-8494-b8a935b6551b
#>  Name: forest-tenure-cut-block-polygons
#> 25: Forest Tenure Managed Licenses Polygons
#>  (oracle_sde)
#>  ID: 04fb4da4-8c6e-4166-a4e8-4c8723443980
#>  Name: forest-tenure-managed-licenses-polygons
#> 26: Forest Tenure Cut Permit Polygons (oracle_sde)
#>  ID: 4876fb83-ce33-4cf8-9c13-7fdeb2f39cf6
#>  Name: forest-tenure-cut-permit-polygons
#> 27: Forest Stewardship Plans - Identified Areas
#>  (multiple, wms, kml, oracle_sde)
#>  ID: dad55255-0f8f-4a4a-a40f-38d22750d065
#>  Name: forest-stewardship-plans-identified-areas
#> 28: Forest Tenure Road Section Lines (multiple, wms,
#>  kml)
#>  ID: 243c94a1-f275-41dc-bc37-91d8a2b26e10
#>  Name: forest-tenure-road-section-lines
#> 29: Forest Tenure Harvesting Authority Polygons
#>  (multiple, wms, kml)
#>  ID: cff7b8f7-6897-444f-8c53-4bb93c7e9f8b
#>  Name: forest-tenure-harvesting-authority-polygons
#> 30: Forest Tenure Real Property Project (multiple,
#>  wms, kml)
#>  ID: 3d7f09c3-ff25-429e-a4d9-d88f9919d904
#>  Name: forest-tenure-real-property-project
#> 31: Forest Tenure Road Section Amendment (multiple)
#>  ID: bad79704-3fc9-43e7-b5d4-660657199af8
#>  Name: forest-tenure-road-section-amendment
#> 32: Forest Tenure Map Notation Point (multiple, wms)
#>  ID: 69f4f070-d70a-43c8-b224-354758f4f492
#>  Name: forest-tenure-map-notation-point
#> 33: Forest Tenure Map Notation Polygon (multiple,
#>  wms, kml)
#>  ID: 78838b54-f241-4269-aeb4-126d14bb5f37
#>  Name: forest-tenure-map-notation-polygon
#> 34: Forest Tenure Map Notation Line (multiple, wms,
#>  kml)
#>  ID: 8ef18f74-e53f-497a-968d-86e9a20419a2
#>  Name: forest-tenure-map-notation-line
#> 35: Forest Tenure Free Use Permit (multiple, wms,
#>  kml)
#>  ID: 814e0ed9-95c3-4750-8152-1307795986f1
#>  Name: forest-tenure-free-use-permit
#> 36: Forest Tenure Communication Site Polygon
#>  (oracle_sde)
#>  ID: d8f0798b-e2e3-48e5-b587-d193daced09e
#>  Name: forest-tenure-communication-site-polygon
#> 37: Forest Tenure Road Segment Lines (multiple, wms,
#>  kml, other)
#>  ID: 9e5bfa62-2339-445e-bf67-81657180c682
#>  Name: forest-tenure-road-segment-lines
#> 38: Forest Tenure Timber Licence Elimination
#>  (multiple, wms, kml)
#>  ID: e15e0543-830f-43b8-b824-ed9a8ed7eeca
#>  Name: forest-tenure-timber-licence-elimination
#> 39: Forest Tenure Timber Licence Remaining (multiple)
#>  ID: edf864e8-bae6-4735-a85b-3e6f276d4a14
#>  Name: forest-tenure-timber-licence-remaining
#> 40: Forest Tenure Road Section Polygon (oracle_sde)
#>  ID: 6c560de6-f322-491f-965a-f75efae6e6d3
#>  Name: forest-tenure-road-section-polygon
#> 41: Forest Tenure Timber License Elimination
#>  (oracle_sde)
#>  ID: d3e5f900-bd2c-4537-8c72-96ee71b4d7ce
#>  Name: forest-tenure-timber-license-elimination
#> 42: Forest Tenure Road Use Permits (multiple, wms,
#>  kml, oracle_sde)
#>  ID: 5c7bc316-1524-4e3f-b799-f8dbafa7a53a
#>  Name: forest-tenure-road-use-permits
#> 43: Forest Operations Map (FOM) - Cutblocks
#>  (multiple)
#>  ID: 7dda4615-5d32-427e-a303-1dcdb90a6fea
#>  Name: forest-operations-map-fom-cutblocks
#> 44: Fertilization Forest Carbon Initiative Projects
#>  (arcgis_rest, other)
#>  ID: 6dec4d9c-f960-4c49-a6df-d4568531db7d
#>  Name:
#>   fertilization-forest-carbon-initiative-projects
#> 45: Forest Carbon Initiative Projects Portal (other)
#>  ID: e861e5eb-5a7f-4f1f-8764-a9b54000767f
#>  Name: forest-carbon-initiative-projects-portal
#> 46: Other Forest Carbon Initiative Projects
#>  (arcgis_rest, other)
#>  ID: 0bdd1284-16f6-4fe5-b451-bd547bf49b67
#>  Name: other-forest-carbon-initiative-projects
#> 47: Reforestation Forest Carbon Initiative Projects
#>  (arcgis_rest, other)
#>  ID: a4f9a983-bc5c-4165-a673-797ded40c9f1
#>  Name:
#>   reforestation-forest-carbon-initiative-projects
#> 48: Lake Classification Vernon Forest District
#>  (oracle_sde)
#>  ID: 9fe70228-f96e-4c2b-9fda-4b08dd443d64
#>  Name: lake-classification-vernon-forest-district
#> 49: Lake Classification Penticton Forest District
#>  (oracle_sde)
#>  ID: 14ffefb0-ef97-499e-abfd-3375ed83ef62
#>  Name: lake-classification-penticton-forest-district
#> 50: Grasslands Fort Nelson Forest District
#>  (oracle_sde)
#>  ID: 91ae918d-5b16-4256-aab6-69de8d769045
#>  Name: grasslands-fort-nelson-forest-district
#> 
#> Access a single record by calling
#>  `bcdc_get_record(ID)` with the ID from the desired
#>  record.

try(
  bcdc_search("regional district", res_format = "fgdb")
)
#> List of B.C. Data Catalogue Records
#> Number of records: 43
#> Titles:
#> 1: Multiple Resource Value Assessment - Coast
#>  Mountains District - Skeena Region - Web Viewer
#>  (other)
#>  ID: c23f158f-193e-444f-baaf-f21a48c1c58b
#>  Name:
#>   multiple-resource-value-assessment-coast-mountains-district-skeena-region-web-viewer
#> 2: Multiple Resource Value Assessment - Sunshine
#>  Coast District - South Coast Region - Web Viewer
#>  (other)
#>  ID: ad20cdab-c8bf-43b4-b393-00fdbab075ac
#>  Name:
#>   multiple-resource-value-assessment-sunshine-coast-district-south-coast-region-web-viewer
#> 3: Integrated Monitoring and Assessment Report -
#>  District of Quesnel - Cariboo Region - Web Viewer
#>  (other)
#>  ID: 50933ec6-e160-4c21-959e-c2ba7821a07a
#>  Name:
#>   integrated-monitoring-and-assessment-report-district-of-quesnel-cariboo-region-web-viewer
#> 4: Recreation Trails for the Cariboo Region (other)
#>  ID: e919372d-5fa1-4214-9f57-1d606a06f1f4
#>  Name: recreation-trails-for-the-cariboo-region
#> 5: Recreation Sites and Trail BC Region Boundaries
#>  (fgdb, multiple, wms, kml)
#>  ID: 919f236a-e40f-4f07-a459-f976f02b6e11
#>  Name:
#>   recreation-sites-and-trail-bc-region-boundaries
#> 6: Cariboo Region Wildfire Recovery Map (BC Map Hub)
#>  (other)
#>  ID: 655e14f0-0de9-4b46-8acc-69a15156fbec
#>  Name:
#>   cariboo-region-wildfire-recovery-map-bc-map-hub-
#> 7: Grassland Benchmark for the Cariboo Region (other)
#>  ID: 2285c49d-a11d-4932-97d4-8c68695e03bb
#>  Name: grassland-benchmark-for-the-cariboo-region
#> 8: Recreation Sites and Trails BC District Boundaries
#>  (fgdb, multiple, wms, kml)
#>  ID: 0ad49841-a25b-4be4-b98c-9e817e94cc48
#>  Name:
#>   recreation-sites-and-trails-bc-district-boundaries
#> 9: Forest Service Road Conditions - Cariboo Natural
#>  Resource Region (other)
#>  ID: b53fd985-5524-4cbe-b179-433c04411dd4
#>  Name:
#>   forest-service-road-conditions-cariboo-natural-resource-region
#> 10: Prince George District Road Alerts Web
#>  Application (other)
#>  ID: c7f2e1e6-84f4-463a-bc16-a07d24da2cf7
#>  Name:
#>   prince-george-district-road-alerts-web-application
#> 11: Quesnel Natural Resource District Small Scale
#>  Salvage Opportunities (other)
#>  ID: 0a68944c-b1b7-42cc-9f12-09965dc327dc
#>  Name:
#>   quesnel-natural-resource-district-small-scale-salvage-opportunities
#> 12: Backcountry Recreation Areas for the Cariboo
#>  Region (other)
#>  ID: 95aaaa75-14a1-4afb-a5cd-e6b5bc8f2dd2
#>  Name:
#>   backcountry-recreation-areas-for-the-cariboo-region
#> 13: High Elevation Viewpoints for the Cariboo Region
#>  (other)
#>  ID: 76e34f98-55f7-461c-9843-30bfe7b72b95
#>  Name:
#>   high-elevation-viewpoints-for-the-cariboo-region
#> 14: Cariboo-Chilcotin Natural Resource District Small
#>  Scale Salvage Opportunities (other)
#>  ID: 45004465-bff3-4601-a355-95e94bb9a8d9
#>  Name:
#>   cariboo-chilcotin-natural-resource-district-small-scale-salvage-opportunities
#> 15: Forest Service Road Conditions Web Map - Cariboo
#>  Natural Resource Region (other)
#>  ID: da349704-60d4-48d0-86ec-0122bc1c8163
#>  Name:
#>   forest-service-road-conditions-web-map-cariboo-natural-resource-region
#> 16: Recreation Trails with Buffer for the Cariboo
#>  Region (other)
#>  ID: e403b4f8-1e78-4a8b-ac05-3d8c305d9f55
#>  Name:
#>   recreation-trails-with-buffer-for-the-cariboo-region
#> 17: Seral Stage Analysis 2004 Cariboo Region - Maps
#>  and Reports (other)
#>  ID: a4fd5e98-804c-4f95-888f-bedcc9964dbd
#>  Name:
#>   seral-stage-analysis-2004-cariboo-region-maps-and-reports
#> 18: Candidate Goal 2 Protected Areas for the Cariboo
#>  Region (other)
#>  ID: e67c7336-d9ac-4ca8-91ed-a512746a517e
#>  Name:
#>   candidate-goal-2-protected-areas-for-the-cariboo-region
#> 19: Timing Windows for Work in and About Waterbodies
#>  in the Cariboo Natural Resource Region (other,
#>  multiple, wms, kml)
#>  ID: fc586323-4f8d-4d64-a178-a74b254131bc
#>  Name:
#>   timing-windows-for-work-in-and-about-waterbodies-in-the-cariboo-natural-resource-region
#> 20: Timing Windows for Work in and About Waterbodies
#>  in the Cariboo NR Region Generalized View (other)
#>  ID: 96a968f8-f9bb-4ff3-9109-1b523706eeef
#>  Name:
#>   timing-windows-for-work-in-and-about-waterbodies-in-the-cariboo-nr-region-generalized-view
#> 21: Timing Windows for Work in and About Waterbodies
#>  in the Cariboo Natural Resource Region - Web Map
#>  (other)
#>  ID: a975db21-3f5f-443c-b2b3-c6e30d37073e
#>  Name:
#>   timing-windows-for-work-in-and-about-waterbodies-in-the-cariboo-natural-resource-region-web-map
#> 22: British Columbia Local Administrative Boundaries
#>  Portal (other)
#>  ID: 8d400574-f0cf-4a67-afbb-4b5be509a537
#>  Name:
#>   british-columbia-local-administrative-boundaries-portal
#> 23: PreparedBC Hazard Map (other)
#>  ID: 6abe8c84-3383-4d2a-84a3-df0a0b3bf405
#>  Name: preparedbc-hazard-map
#> 24: Forest Service Road Safety Information Web Map
#>  Application - South Coast (other)
#>  ID: 55af70f9-a4f9-4468-8aca-4ba727969ec4
#>  Name:
#>   forest-service-road-safety-information-web-map-application-south-coast
#> 25: Haida Gwaii - Road and Safety Web App (other,
#>  arcgis_rest)
#>  ID: 2f772815-efdd-4829-b684-5fd17c73ce9b
#>  Name: haida-gwaii-road-and-safety-web-app
#> 26: Forest Service Road Safety and Closure
#>  Information - South Coast (other)
#>  ID: 018ed540-24a4-4b7f-9dfd-ce837f0d812f
#>  Name:
#>   forest-service-road-safety-and-closure-information-south-coast
#> 27: Timber Harvesting Land Base - Data - Howe Sound
#>  Cumulative Effects (other)
#>  ID: 9a2c42ce-7d51-4ace-a065-48a00ec0b804
#>  Name:
#>   timber-harvesting-land-base-data-howe-sound-cumulative-effects
#> 28: CMHC Primary Rental Market (BC) (other, pdf)
#>  ID: aa7a5049-4a47-4182-b2c5-f4737f1913e0
#>  Name: cmhc-primary-rental-market-bc-
#> 29: Legal Planning Objectives - All - Polygon
#>  (multiple, other)
#>  ID: eed2b09d-295a-4a02-adf2-178283ffacb7
#>  Name: legal-planning-objectives-all-polygon
#> 30: Custom Census Reports (2021, 2016, 2011, 2006)
#>  (other, xlsx)
#>  ID: 894497d9-6901-4ed8-8655-79010e4b9b09
#>  Name: custom-census-reports-2021-2016-2011-2006-
#> 31: Legal Planning Objectives - Current - Polygon
#>  (multiple, wms, kml, other)
#>  ID: 2c02040c-d7c5-4960-8d04-dea01d6d3e9f
#>  Name: legal-planning-objectives-current-polygon
#> 32: Local Government Tax Rates and Assessments 2013
#>  (other)
#>  ID: 93d9cdb1-f28a-46a5-8f16-b4d900bdcd18
#>  Name:
#>   local-government-tax-rates-and-assessments-2013
#> 33: Local Government Tax Rates and Assessment 2014
#>  (other)
#>  ID: 7fcbc227-7406-4bd9-aabf-7086ff445535
#>  Name: local-government-tax-rates-and-assessment-2014
#> 34: Strategic Land and Resource Plans - All
#>  (multiple, other)
#>  ID: 298d1034-c1be-4fd1-ad4b-d00ad5ab4b88
#>  Name: strategic-land-and-resource-plans-all
#> 35: Old Growth - Web Viewer - South Coast SBOT
#>  (other)
#>  ID: d6309f84-d1e0-45ea-afff-6a3293825ce2
#>  Name: old-growth-web-viewer-south-coast-sbot
#> 36: BC Wildfire Community Wildfire Protection Plan
#>  (CWPP) Locations - Internal (other)
#>  ID: 513e1310-d2c4-486f-8923-3b06caff2dc7
#>  Name:
#>   bc-wildfire-community-wildfire-protection-plan-locations-internal
#> 37: Strategic Land and Resource Plans - Current
#>  (multiple, wms, kml, other, oracle_sde)
#>  ID: 4b142d4c-83d6-4ecc-b66c-66601ae65992
#>  Name: strategic-land-and-resource-plans-current
#> 38: Soil Relocation Receiving Sites (arcgis_rest,
#>  other)
#>  ID: 549cb689-b602-45a8-b56c-49726891cd65
#>  Name: soil-relocation-receiving-sites
#> 39: Soil Relocation Source Sites (arcgis_rest, other)
#>  ID: 404dfaf6-7c94-4b34-86a5-8d445f19ffbd
#>  Name: soil-relocation-source-sites
#> 40: High Volume Receiving Sites (arcgis_rest, other)
#>  ID: 5bc1cbef-c802-4dc6-af87-b5813b3a4a77
#>  Name: high-volume-receiving-sites
#> 41: Non Legal Planning Features - All - Polygon
#>  (multiple, other)
#>  ID: ad71d3c2-0776-4233-9765-18fbf18580cf
#>  Name: non-legal-planning-features-all-polygon
#> 42: Old Growth - BC Cumulative Effects Framework -
#>  2019 Assessment - Robson Valley (fgdb, multiple)
#>  ID: 502442fe-0d45-43f3-a8e5-f733e5c74d7d
#>  Name:
#>   old-growth-bc-cumulative-effects-framework-2019-assessment-robson-valley
#> 43: Non Legal Planning Features - Current - Polygon
#>  (multiple, wms, kml, other)
#>  ID: 5d859a89-f173-4006-82f9-16254de2c1fc
#>  Name: non-legal-planning-features-current-polygon
#> 
#> Access a single record by calling
#>  `bcdc_get_record(ID)` with the ID from the desired
#>  record.

try(
  bcdc_search("angling", groups = "bc-tourism")
)
#> List of B.C. Data Catalogue Records
#> Number of records: 1
#> Titles:
#> 1: Angling Licence Sales Statistics 2010 to Current
#>  (xlsx, csv)
#>  ID: ebbe3328-43ac-4440-be2d-b3f83ae03780
#>  Name:
#>   angling-licence-sales-statistics-2010-to-current
#> 
#> Access a single record by calling
#>  `bcdc_get_record(ID)` with the ID from the desired
#>  record.
# }