Search the B.C. Data Catalogue

Search the B.C. Data Catalogue

Usage

bcdc_search(
  ...,
  license_id = NULL,
  download_audience = "Public",
  res_format = NULL,
  sector = NULL,
  organization = NULL,
  n = 100
)

Arguments

...

search terms

license_id

the type of license (see bcdc_search_facets("license_id")).

download_audience

download audience (see bcdc_search_facets("download_audience")). Default "Public"

res_format

format of resource (see bcdc_search_facets("res_format"))

sector

sector of government from which the data comes (see bcdc_search_facets("sector"))

organization

government organization that manages the data (see bcdc_search_facets("organization"))

n

number of results to return. Default 100

Value

A list containing the records that match the search

Examples

# \donttest{
try(
  bcdc_search("forest")
)
#> Found 296 matches. Returning the first 100.
#> To see them all, rerun the search and set the 'n' argument to 296.
#> List of B.C. Data Catalogue Records
#> Number of records: 100
#> Showing the top 50 results. You can assign the
#>  output of bcdc_search, to an object and subset with
#>  `[` to see other results in the set.
#> 
#> Titles:
#> 1: Forest Development Units (multiple, wms, kml,
#>  oracle_sde)
#>  ID: 25bbc9a6-f0eb-47dd-849c-3161a6512468
#>  Name: forest-development-units
#> 2: FADM - Provincial Forest (multiple, wms, kml)
#>  ID: 4a023cb2-09fd-44c2-b3ce-439ecb5f39f3
#>  Name: fadm-provincial-forest
#> 3: Forest Inventory Zones (multiple, wms, kml)
#>  ID: 67e95c68-c1ef-4363-b351-0dfead151122
#>  Name: forest-inventory-zones
#> 4: Forest Tenure Managed Licence (multiple, wms, kml)
#>  ID: c3e96239-cdc9-4328-ac19-58fba1623ef8
#>  Name: forest-tenure-managed-licence
#> 5: FDU - Forest Licence Reference (multiple)
#>  ID: 7e986391-c740-4d1b-9294-7576af2d2998
#>  Name: fdu-forest-licence-reference
#> 6: FADM - Provincial Forest Deletion (multiple, wms,
#>  kml)
#>  ID: fc480dec-1e67-4944-9343-6c72ae124440
#>  Name: fadm-provincial-forest-deletion
#> 7: Forest Tenure Communication Sites (multiple, wms,
#>  kml)
#>  ID: 1361d483-26ae-4ca9-adba-e19c3efac147
#>  Name: forest-tenure-communication-sites
#> 8: Forest Tenure Timber Licence (multiple, wms, kml)
#>  ID: f2356651-4456-4d93-ba17-7b8c4a3ee53d
#>  Name: forest-tenure-timber-licence
#> 9: FADM - Provincial Forest Addition (multiple, wms,
#>  kml)
#>  ID: 07dba191-3ebc-4c32-af7e-16769c237011
#>  Name: fadm-provincial-forest-addition
#> 10: FADM - Provincial Forest Exclusion (multiple,
#>  wms, kml)
#>  ID: f3239e0c-f347-49f9-a53d-39290e597159
#>  Name: fadm-provincial-forest-exclusion
#> 11: Forest Road Segment Tenure (multiple, wms, kml)
#>  ID: 07b0830e-a505-4c88-8d9a-8cf0a630f565
#>  Name: forest-road-segment-tenure
#> 12: RESULTS - Forest Cover Silviculture (multiple,
#>  wms, kml)
#>  ID: 258bb088-4113-47b1-b568-ce20bd64e3e3
#>  Name: results-forest-cover-silviculture
#> 13: RESULTS - Forest Cover Inventory (multiple, wms)
#>  ID: 56ac43a7-724a-4f01-b193-d5f9a16ef0a8
#>  Name: results-forest-cover-inventory
#> 14: Generalized Forest Cover Ownership (wms, kml,
#>  multiple)
#>  ID: 5fc4e8ce-dd1d-44fd-af17-e0789cf65e4e
#>  Name: generalized-forest-cover-ownership
#> 15: RESULTS - Forest Cover Reserve (multiple, wms,
#>  kml)
#>  ID: 7028ac47-45dd-41d7-a371-be5a04177afe
#>  Name: results-forest-cover-reserve
#> 16: Forest Licensee Operating Areas (multiple, wms,
#>  kml)
#>  ID: 5a18c401-7c99-48fe-8a1a-b4597072d8b6
#>  Name: forest-licensee-operating-areas
#> 17: Forest Tenure Road Section Lines (multiple, wms,
#>  kml)
#>  ID: 243c94a1-f275-41dc-bc37-91d8a2b26e10
#>  Name: forest-tenure-road-section-lines
#> 18: Forest Tenure Harvesting Authority Polygons
#>  (multiple, wms, kml)
#>  ID: cff7b8f7-6897-444f-8c53-4bb93c7e9f8b
#>  Name: forest-tenure-harvesting-authority-polygons
#> 19: Forest Tenure Real Property Project (multiple,
#>  wms, kml)
#>  ID: 3d7f09c3-ff25-429e-a4d9-d88f9919d904
#>  Name: forest-tenure-real-property-project
#> 20: Forest Tenure Map Notation Point (multiple, wms)
#>  ID: 69f4f070-d70a-43c8-b224-354758f4f492
#>  Name: forest-tenure-map-notation-point
#> 21: Forest Tenure Map Notation Polygon (multiple,
#>  wms, kml)
#>  ID: 78838b54-f241-4269-aeb4-126d14bb5f37
#>  Name: forest-tenure-map-notation-polygon
#> 22: Forest Tenure Map Notation Line (multiple, wms,
#>  kml)
#>  ID: 8ef18f74-e53f-497a-968d-86e9a20419a2
#>  Name: forest-tenure-map-notation-line
#> 23: Forest Tenure Road Section Amendment (multiple)
#>  ID: bad79704-3fc9-43e7-b5d4-660657199af8
#>  Name: forest-tenure-road-section-amendment
#> 24: Forest Tenure Free Use Permit (multiple, wms,
#>  kml)
#>  ID: 814e0ed9-95c3-4750-8152-1307795986f1
#>  Name: forest-tenure-free-use-permit
#> 25: Forest Tenure Road Segment Lines (multiple, wms,
#>  kml)
#>  ID: 9e5bfa62-2339-445e-bf67-81657180c682
#>  Name: forest-tenure-road-segment-lines
#> 26: Forest Tenure Timber Licence Elimination
#>  (multiple, wms, kml)
#>  ID: e15e0543-830f-43b8-b824-ed9a8ed7eeca
#>  Name: forest-tenure-timber-licence-elimination
#> 27: Forest Tenure Timber Licence Remaining (multiple)
#>  ID: edf864e8-bae6-4735-a85b-3e6f276d4a14
#>  Name: forest-tenure-timber-licence-remaining
#> 28: Forest Tenure Road Use Permits (multiple, wms,
#>  kml, oracle_sde)
#>  ID: 5c7bc316-1524-4e3f-b799-f8dbafa7a53a
#>  Name: forest-tenure-road-use-permits
#> 29: Biodiversity Corridors Kamloops Forest District
#>  (multiple)
#>  ID: f15d11d3-2a3e-41a5-a576-dab136cced9b
#>  Name:
#>   biodiversity-corridors-kamloops-forest-district
#> 30: Forest Tenure Special Access Road Line (multiple,
#>  wms, kml)
#>  ID: 4ee37a5d-6a14-4f98-8ad9-566b4ea28e2b
#>  Name: forest-tenure-special-access-road-line
#> 31: Forest Tenure Special Access Road Polygon
#>  (multiple, wms, kml)
#>  ID: 6d171eb8-d7a8-4ad5-8a9e-027e43f7a54b
#>  Name: forest-tenure-special-access-road-polygon
#> 32: Forest Tenure Special Use Permit Polygon
#>  (multiple, wms, kml)
#>  ID: d29b37fc-cbfe-4ebf-ac6e-2bf8fa926a81
#>  Name: forest-tenure-special-use-permit-polygon
#> 33: VRI - 2020 - Forest Vegetation Composite Polygons
#>  (fgdb)
#>  ID: 6ba30649-14cd-44ad-a11f-794feed39f40
#>  Name: vri-2020-forest-vegetation-composite-polygons
#> 34: RESULTS - Forest Cover Species (Attribute Only)
#>  (multiple)
#>  ID: 2db77f8b-765b-488f-9752-f1349cf6ebfc
#>  Name: results-forest-cover-species-attribute-only
#> 35: Potential Moose Habitat Lillooet Forest District
#>  (multiple)
#>  ID: de3be33c-18b1-40e7-a61c-70784ad20a15
#>  Name:
#>   potential-moose-habitat-lillooet-forest-district
#> 36: Forest Tenure Cutblock Polygons (FTA 4.0)
#>  (multiple, wms, kml)
#>  ID: dfb8b498-fa4b-4286-b3ec-58db88aca1cf
#>  Name: forest-tenure-cutblock-polygons-fta-4-0
#> 37: Productive Forest Landbase for the Cariboo Region
#>  (multiple, wms, kml)
#>  ID: 0cfa7f53-272d-47d4-83df-95b4edd75460
#>  Name:
#>   productive-forest-landbase-for-the-cariboo-region
#> 38: Non-Productive Forest for the Cariboo Region
#>  (multiple, wms, kml)
#>  ID: 8f0eabe0-f424-4062-ade1-cb4368c81b3a
#>  Name: non-productive-forest-for-the-cariboo-region
#> 39: North Thompson Caribou Habitat Clearwater Forest
#>  District (multiple)
#>  ID: 6271d5dc-4870-41e4-903d-4d378c297353
#>  Name:
#>   north-thompson-caribou-habitat-clearwater-forest-district
#> 40: Caribou Movement Corridors Clearwater Forest
#>  District - Lines (multiple)
#>  ID: 7c7768f2-7a3d-4a8f-958e-e7a8dd2b18f2
#>  Name:
#>   caribou-movement-corridors-clearwater-forest-district-lines
#> 41: Rose Swanson Sensitive Area Vernon Forest
#>  District (multiple)
#>  ID: 30a88772-f815-4795-8f4d-9c09d52935a9
#>  Name:
#>   rose-swanson-sensitive-area-vernon-forest-district
#> 42: Recreation Sites, Reserves, and Interpretive
#>  Forests Details and Closures (wms, kml, multiple)
#>  ID: b0f23ae8-e0aa-449c-a806-653ba37f100a
#>  Name:
#>   recreation-sites-reserves-and-interpretive-forests-details-and-closures
#> 43: Mixed Forest 1:250,000 GeoBase Land Cover
#>  (multiple)
#>  ID: 902f5af2-9fce-4579-b48d-8bd7f85d36dd
#>  Name: mixed-forest-1-250-000-geobase-land-cover
#> 44: Coniferous Forest 1:250,000 GeoBase Land Cover
#>  (multiple)
#>  ID: 177b5ef0-8339-4fd7-a40a-93edb3797d07
#>  Name: coniferous-forest-1-250-000-geobase-land-cover
#> 45: Broadleaf Forest 1:250,000 GeoBase Land Cover
#>  (multiple)
#>  ID: 78bb0ee9-96ed-49fb-9c0c-7434b2f24d7e
#>  Name: broadleaf-forest-1-250-000-geobase-land-cover
#> 46: Old Growth Technical Advisory Panel (TAP) -
#>  Recruitment Forest (pdf, multiple, oracle_sde)
#>  ID: d866c00d-c9d5-4af9-a171-a797f524e401
#>  Name:
#>   old-growth-technical-advisory-panel-tap-recruitment-forest
#> 47: Old Growth Technical Advisory Panel (TAP) -
#>  Ancient Forest (pdf, multiple, other, wms, kml)
#>  ID: b43bc442-8a0f-42f2-b085-04f4771b2f28
#>  Name:
#>   old-growth-technical-advisory-panel-tap-ancient-forest
#> 48: Old Growth Technical Advisory Panel (TAP) - Old
#>  Forests (pdf, multiple)
#>  ID: 6f08c9d8-d6c0-4f08-ac64-54585c92ccb0
#>  Name:
#>   old-growth-technical-advisory-panel-tap-old-forests
#> 49: Purchasing Card Expenditures - Forest Practices
#>  Board - FYE11 (csv)
#>  ID: 618260d9-5dfe-4d98-805b-f11d9b4a94e1
#>  Name:
#>   purchasing-card-expenditures-forest-practices-board-fye11
#> 50: Forest Visual Quality - Data - Howe Sound
#>  Cumulative Effects (fgdb, geojson, csv)
#>  ID: ea0662a3-855d-403e-9826-48c66d5db348
#>  Name:
#>   forest-visual-quality-data-howe-sound-cumulative-effects
#> 
#> Access a single record by calling
#>  `bcdc_get_record(ID)` with the ID from the desired
#>  record.

try(
  bcdc_search("regional district", res_format = "fgdb")
)
#> List of B.C. Data Catalogue Records
#> Number of records: 16
#> Titles:
#> 1: Recreation Trails for the Cariboo Region (other)
#>  ID: e919372d-5fa1-4214-9f57-1d606a06f1f4
#>  Name: recreation-trails-for-the-cariboo-region
#> 2: Recreation Sites and Trail BC Region Boundaries
#>  (fgdb, multiple, wms, kml)
#>  ID: 919f236a-e40f-4f07-a459-f976f02b6e11
#>  Name:
#>   recreation-sites-and-trail-bc-region-boundaries
#> 3: Grassland Benchmark for the Cariboo Region (other)
#>  ID: 2285c49d-a11d-4932-97d4-8c68695e03bb
#>  Name: grassland-benchmark-for-the-cariboo-region
#> 4: Recreation Sites and Trails BC District Boundaries
#>  (fgdb, multiple, wms, kml)
#>  ID: 0ad49841-a25b-4be4-b98c-9e817e94cc48
#>  Name:
#>   recreation-sites-and-trails-bc-district-boundaries
#> 5: Backcountry Recreation Areas for the Cariboo
#>  Region (other)
#>  ID: 95aaaa75-14a1-4afb-a5cd-e6b5bc8f2dd2
#>  Name:
#>   backcountry-recreation-areas-for-the-cariboo-region
#> 6: High Elevation Viewpoints for the Cariboo Region
#>  (other)
#>  ID: 76e34f98-55f7-461c-9843-30bfe7b72b95
#>  Name:
#>   high-elevation-viewpoints-for-the-cariboo-region
#> 7: Recreation Trails with Buffer for the Cariboo
#>  Region (other)
#>  ID: e403b4f8-1e78-4a8b-ac05-3d8c305d9f55
#>  Name:
#>   recreation-trails-with-buffer-for-the-cariboo-region
#> 8: Seral Stage Analysis 2004 Cariboo Region - Maps
#>  and Reports (other)
#>  ID: a4fd5e98-804c-4f95-888f-bedcc9964dbd
#>  Name:
#>   seral-stage-analysis-2004-cariboo-region-maps-and-reports
#> 9: Candidate Goal 2 Protected Areas for the Cariboo
#>  Region (other)
#>  ID: e67c7336-d9ac-4ca8-91ed-a512746a517e
#>  Name:
#>   candidate-goal-2-protected-areas-for-the-cariboo-region
#> 10: Custom Census Reports (2016, 2011, 2006) (other,
#>  xlsx)
#>  ID: 894497d9-6901-4ed8-8655-79010e4b9b09
#>  Name: custom-census-reports-2016-2011-2006-
#> 11: CMHC Primary Rental Market (BC) (other, pdf)
#>  ID: aa7a5049-4a47-4182-b2c5-f4737f1913e0
#>  Name: cmhc-primary-rental-market-bc-
#> 12: Legal Planning Objectives - Current - Polygon
#>  (multiple, wms, kml, other)
#>  ID: 2c02040c-d7c5-4960-8d04-dea01d6d3e9f
#>  Name: legal-planning-objectives-current-polygon
#> 13: Local Government Tax Rates and Assessments 2013
#>  (other)
#>  ID: 93d9cdb1-f28a-46a5-8f16-b4d900bdcd18
#>  Name:
#>   local-government-tax-rates-and-assessments-2013
#> 14: Local Government Tax Rates and Assessment 2014
#>  (other)
#>  ID: 7fcbc227-7406-4bd9-aabf-7086ff445535
#>  Name: local-government-tax-rates-and-assessment-2014
#> 15: Strategic Land and Resource Plans - Current
#>  (multiple, wms, kml, other, oracle_sde)
#>  ID: 4b142d4c-83d6-4ecc-b66c-66601ae65992
#>  Name: strategic-land-and-resource-plans-current
#> 16: Non Legal Planning Features - Current - Polygon
#>  (multiple, wms, kml, other)
#>  ID: 5d859a89-f173-4006-82f9-16254de2c1fc
#>  Name: non-legal-planning-features-current-polygon
#> 
#> Access a single record by calling
#>  `bcdc_get_record(ID)` with the ID from the desired
#>  record.
# }