Search the B.C. Data Catalogue

Search the B.C. Data Catalogue

bcdc_search(
  ...,
  license_id = NULL,
  download_audience = "Public",
  type = NULL,
  res_format = NULL,
  sector = NULL,
  organization = NULL,
  n = 100
)

Arguments

...
search terms

license_id
the type of license (see bcdc_search_facets("license_id")).

download_audience
download audience (see bcdc_search_facets("download_audience")). Default "Public"

type
type of resource (see bcdc_search_facets("type"))

res_format
format of resource (see bcdc_search_facets("res_format"))

sector
sector of government from which the data comes (see bcdc_search_facets("sector"))

organization
government organization that manages the data (see bcdc_search_facets("organization"))

n
number of results to return. Default 100

Value

A list containing the records that match the search

Examples

# \donttest{
bcdc_search("forest")

#> Found 305 matches. Returning the first 100.
#> To see them all, rerun the search and set the 'n' argument to 305.

#> List of B.C. Data Catalogue Records
#> Number of records: 100
#>  (Showing the top 10)
#> Titles:
#> 1: RESULTS - Forest Cover Inventory (other, wms)
#>  ID: 56ac43a7-724a-4f01-b193-d5f9a16ef0a8
#>  Name: results-forest-cover-inventory
#> 2: RESULTS - Forest Cover Silviculture (other, wms, kml)
#>  ID: 258bb088-4113-47b1-b568-ce20bd64e3e3
#>  Name: results-forest-cover-silviculture
#> 3: FDU - Forest Licence Reference (other)
#>  ID: 7e986391-c740-4d1b-9294-7576af2d2998
#>  Name: fdu-forest-licence-reference
#> 4: RESULTS - Forest Cover Species (Attribute Only) (other)
#>  ID: 2db77f8b-765b-488f-9752-f1349cf6ebfc
#>  Name: results-forest-cover-species-attribute-only
#> 5: Forest Tenure Managed Licence (other, wms, kml)
#>  ID: c3e96239-cdc9-4328-ac19-58fba1623ef8
#>  Name: forest-tenure-managed-licence
#> 6: Forest Inventory Zones (other, wms, kml)
#>  ID: 67e95c68-c1ef-4363-b351-0dfead151122
#>  Name: forest-inventory-zones
#> 7: Harvested Areas of BC (Consolidated Cutblocks) (other, fgdb, wms,
#>  kml, pdf)
#>  ID: b1b647a6-f271-42e0-9cd0-89ec24bce9f7
#>  Name: harvested-areas-of-bc-consolidated-cutblocks-
#> 8: Mixed Forest 1:250,000 GeoBase Land Cover (other)
#>  ID: 902f5af2-9fce-4579-b48d-8bd7f85d36dd
#>  Name: mixed-forest-1-250-000-geobase-land-cover
#> 9: Coniferous Forest 1:250,000 GeoBase Land Cover (other)
#>  ID: 177b5ef0-8339-4fd7-a40a-93edb3797d07
#>  Name: coniferous-forest-1-250-000-geobase-land-cover
#> 10: Broadleaf Forest 1:250,000 GeoBase Land Cover (other)
#>  ID: 78bb0ee9-96ed-49fb-9c0c-7434b2f24d7e
#>  Name: broadleaf-forest-1-250-000-geobase-land-cover
#> 
#> Access a single record by calling `bcdc_get_record(ID)` with the ID
#>  from the desired record.

bcdc_search("regional district", type = "Geographic", res_format = "fgdb")

#> Found 190 matches. Returning the first 100.
#> To see them all, rerun the search and set the 'n' argument to 190.

#> List of B.C. Data Catalogue Records
#> Number of records: 100
#>  (Showing the top 10)
#> Titles:
#> 1: Regional district financial statistics (other)
#>  ID: 35fd9fe7-e937-4719-82e8-852cc699a113
#>  Name: regional-district-financial-statistics
#> 2: Natural Resource (NR) Districts (other, wms)
#>  ID: 0bc73892-e41f-41d0-8d8e-828c16139337
#>  Name: natural-resource-nr-district
#> 3: Recreation Sites and Trails BC District Boundaries (fgdb, other,
#>  wms, kml)
#>  ID: 0ad49841-a25b-4be4-b98c-9e817e94cc48
#>  Name: recreation-sites-and-trails-bc-district-boundaries
#> 4: Regional District Boundaries - Road centreline aligned (other, wms,
#>  kml)
#>  ID: b66ed992-16c9-4527-b363-0ae4827d103b
#>  Name: regional-district-boundaries-road-centreline-aligned
#> 5: Regional Districts - Legally Defined Administrative Areas of BC
#>  (other, xlsx, wms, kml)
#>  ID: d1aff64e-dbfe-45a6-af97-582b7f6418b9
#>  Name: regional-districts-legally-defined-administrative-areas-of-bc
#> 6: Legally Defined Administrative Areas of BC (other, xlsx)
#>  ID: 12f0fd64-94d5-422a-832e-6627a2a01702
#>  Name: legally-defined-administrative-areas-of-bc
#> 7: Local and Regional Greenspaces (wms, kml, xlsx, other)
#>  ID: 6a2fea1b-0cc4-4fc2-8017-eaf755d516da
#>  Name: local-and-regional-greenspaces
#> 8: BC College Region Boundaries (other, wms, kml)
#>  ID: 6b8da230-9bc2-4bb4-989e-7d5dad84a4ac
#>  Name: bc-college-region-boundaries
#> 9: Communities Affected by Evacuation Orders and Alerts (other, wms,
#>  kml)
#>  ID: 284f6b3f-e1ce-4394-81aa-1a4af9feee56
#>  Name: communities-affected-by-evacuation-orders-and-alerts
#> 10: Recreation Sites and Trail BC Region Boundaries (fgdb, other, wms,
#>  kml)
#>  ID: 919f236a-e40f-4f07-a459-f976f02b6e11
#>  Name: recreation-sites-and-trail-bc-region-boundaries
#> 
#> Access a single record by calling `bcdc_get_record(ID)` with the ID
#>  from the desired record.

# }