as_tibble

See tibble::as_tibble for details.

After tuning a query, collect() is used to actually bring the data into memory. This will retrieve an sf object into R. The as_tibble() function can be used interchangeably with collect which matches dbplyr behaviour.

See dplyr::collect for details.

# S3 method for bcdc_promise
collect(x, ...)

# S3 method for bcdc_promise
as_tibble(x, ...)

Arguments

x
object of class bcdc_promise

Examples

# \donttest{
try(
  bcdc_query_geodata("bc-airports") %>%
    collect()
)

#> Simple feature collection with 455 features and 41 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 406543.7 ymin: 367957.6 xmax: 1796645 ymax: 1689146
#> Projected CRS: NAD83 / BC Albers
#> # A tibble: 455 x 42
#>    id      CUSTODIAN_ORG_DES… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>  * <chr>   <chr>              <chr>            <chr>            <chr>           
#>  1 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  2 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  3 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  4 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  5 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  6 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  7 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  8 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  9 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 10 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> # … with 445 more rows, and 37 more variables: SOURCE_DATA_ID <chr>,
#> #   SUPPLIED_SOURCE_ID_IND <chr>, AIRPORT_NAME <chr>, DESCRIPTION <chr>,
#> #   PHYSICAL_ADDRESS <chr>, ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>,
#> #   POSTAL_CODE <chr>, LOCALITY <chr>, CONTACT_PHONE <chr>,
#> #   CONTACT_EMAIL <chr>, CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>,
#> #   LATITUDE <dbl>, LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <dbl>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

try(
  bcdc_query_geodata("bc-airports") %>%
    as_tibble()
)

#> Simple feature collection with 455 features and 41 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 406543.7 ymin: 367957.6 xmax: 1796645 ymax: 1689146
#> Projected CRS: NAD83 / BC Albers
#> # A tibble: 455 x 42
#>    id      CUSTODIAN_ORG_DES… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>  * <chr>   <chr>              <chr>            <chr>            <chr>           
#>  1 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  2 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  3 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  4 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  5 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  6 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  7 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  8 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#>  9 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 10 WHSE_I… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> # … with 445 more rows, and 37 more variables: SOURCE_DATA_ID <chr>,
#> #   SUPPLIED_SOURCE_ID_IND <chr>, AIRPORT_NAME <chr>, DESCRIPTION <chr>,
#> #   PHYSICAL_ADDRESS <chr>, ALIAS_ADDRESS <chr>, STREET_ADDRESS <chr>,
#> #   POSTAL_CODE <chr>, LOCALITY <chr>, CONTACT_PHONE <chr>,
#> #   CONTACT_EMAIL <chr>, CONTACT_FAX <chr>, WEBSITE_URL <chr>, IMAGE_URL <chr>,
#> #   LATITUDE <dbl>, LONGITUDE <dbl>, KEYWORDS <chr>, DATE_UPDATED <date>,
#> #   SITE_GEOCODED_IND <chr>, AERODROME_STATUS <chr>, AIRCRAFT_ACCESS_IND <chr>,
#> #   DATA_SOURCE <chr>, DATA_SOURCE_YEAR <chr>, ELEVATION <dbl>,
#> #   FUEL_AVAILABILITY_IND <chr>, HELICOPTER_ACCESS_IND <chr>, IATA_CODE <chr>,
#> #   ICAO_CODE <chr>, MAX_RUNWAY_LENGTH <dbl>, NUMBER_OF_RUNWAYS <int>,
#> #   OIL_AVAILABILITY_IND <chr>, RUNWAY_SURFACE <chr>,
#> #   SEAPLANE_ACCESS_IND <chr>, TC_LID_CODE <chr>, SEQUENCE_ID <int>,
#> #   SE_ANNO_CAD_DATA <chr>, geometry <POINT [m]>

# }