Filter a query from Web Service call

Filter a query from Web Service using dplyr methods. This filtering is accomplished lazily so that the full sf object is not read into memory until collect() has been called.

See dplyr::filter for details.

# S3 method for bcdc_promise
filter(.data, ...)

Arguments

.data
object of class bcdc_promise (likely passed from bcdc_query_geodata())

...
Logical predicates with which to filter the results. Multiple conditions are combined with &. Only rows where the condition evaluates to TRUE are kept. Accepts normal R expressions as well as any of the special CQL geometry functions such as WITHIN() or INTERSECTS(). If you know CQL and want to write a CQL query directly, write it enclosed in quotes, wrapped in the CQL() function. e.g., CQL("ID = '42'")

Methods (by class)

bcdc_promise: filter.bcdc_promise

Examples

# \donttest{
try(
  crd <- bcdc_query_geodata("regional-districts-legally-defined-administrative-areas-of-bc") %>%
    filter(ADMIN_AREA_NAME == "Cariboo Regional District") %>%
    collect()
)

try(
  ret1 <- bcdc_query_geodata("fire-perimeters-historical") %>%
    filter(FIRE_YEAR == 2000, FIRE_CAUSE == "Person", INTERSECTS(crd)) %>%
    collect()
)
  # }