Select columns from Web Service call

Similar to a dplyr::select call, this allows you to select which columns you want the Web Service to return. A key difference between dplyr::select and bcdata::select is the presence of "sticky" columns that are returned regardless of what columns are selected. If any of these "sticky" columns are selected only "sticky" columns are return. bcdc_describe_feature is one way to tell if columns are sticky in advance of issuing the Web Service call.

See dplyr::select for details.

# S3 method for bcdc_promise
select(.data, ...)

Arguments

.data
object of class bcdc_promise (likely passed from bcdc_query_geodata())

...
One or more unquoted expressions separated by commas. See details.

Methods (by class)

bcdc_promise: select.bcdc_promise

Examples

# \donttest{
try(
  feature_spec <- bcdc_describe_feature("bc-airports")
)

try(
  ## Columns that can selected:
  feature_spec[feature_spec$sticky == TRUE,]
)

#> # A tibble: 33 x 4
#>    col_name         sticky remote_col_type local_col_type
#>    <chr>            <lgl>  <chr>           <chr>         
#>  1 DESCRIPTION      TRUE   xsd:string      character     
#>  2 PHYSICAL_ADDRESS TRUE   xsd:string      character     
#>  3 ALIAS_ADDRESS    TRUE   xsd:string      character     
#>  4 STREET_ADDRESS   TRUE   xsd:string      character     
#>  5 POSTAL_CODE      TRUE   xsd:string      character     
#>  6 LOCALITY         TRUE   xsd:string      character     
#>  7 CONTACT_PHONE    TRUE   xsd:string      character     
#>  8 CONTACT_EMAIL    TRUE   xsd:string      character     
#>  9 CONTACT_FAX      TRUE   xsd:string      character     
#> 10 WEBSITE_URL      TRUE   xsd:string      character     
#> # … with 23 more rows

## Select columns
try(
  bcdc_query_geodata("bc-airports") %>%
    select(DESCRIPTION, PHYSICAL_ADDRESS)
)

#> Querying 'bc-airports' record
#> ● Using collect() on this object will return 455
#> ● features and 11 fields
#> ● At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 11 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 833323.9 ymin: 381604.1 xmax: 1198292 ymax: 1054950
#> projected CRS:  NAD83 / BC Albers
#> # A tibble: 6 x 12
#>   id       CUSTODIAN_ORG_DES… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>   <chr>    <chr>              <chr>            <chr>            <chr>           
#> 1 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 2 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 3 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 4 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 5 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 6 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> # … with 7 more variables: SOURCE_DATA_ID <chr>, SUPPLIED_SOURCE_ID_IND <chr>,
#> #   AIRPORT_NAME <chr>, DESCRIPTION <chr>, PHYSICAL_ADDRESS <chr>,
#> #   SEQUENCE_ID <int>, geometry <POINT [m]>

## Select "sticky" columns
try(
  bcdc_query_geodata("bc-airports") %>%
    select(LOCALITY)
)

#> Querying 'bc-airports' record
#> ● Using collect() on this object will return 455
#> ● features and 10 fields
#> ● At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 10 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 833323.9 ymin: 381604.1 xmax: 1198292 ymax: 1054950
#> projected CRS:  NAD83 / BC Albers
#> # A tibble: 6 x 11
#>   id       CUSTODIAN_ORG_DES… BUSINESS_CATEGO… BUSINESS_CATEGO… OCCUPANT_TYPE_D…
#>   <chr>    <chr>              <chr>            <chr>            <chr>           
#> 1 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 2 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 3 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 4 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 5 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> 6 WHSE_IM… "Ministry of Fore… airTransportati… Air Transportat… BC Airports     
#> # … with 6 more variables: SOURCE_DATA_ID <chr>, SUPPLIED_SOURCE_ID_IND <chr>,
#> #   AIRPORT_NAME <chr>, LOCALITY <chr>, SEQUENCE_ID <int>, geometry <POINT [m]>

# }